% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae_accessors.R
\name{accessors}
\alias{accessors}
\alias{get_adjacencyMatrix}
\alias{get_adjacencyMatrix,SummarizedExperiment-method}
\alias{get_group}
\alias{get_group,MultiAssayExperiment-method}
\alias{get_metacell}
\alias{get_metacell,SummarizedExperiment-method}
\alias{get_cc}
\alias{get_cc,SummarizedExperiment-method}
\alias{get_proteinID}
\alias{get_proteinID,SummarizedExperiment-method}
\alias{get_colID}
\alias{get_colID,SummarizedExperiment-method}
\alias{get_type}
\alias{get_type,SummarizedExperiment-method}
\alias{get_pkg_version}
\alias{get_pkg_version,SummarizedExperiment-method}
\title{Accessors functions}
\usage{
get_adjacencyMatrix(object, ...)

\S4method{get_adjacencyMatrix}{SummarizedExperiment}(object)

get_group(object, ...)

\S4method{get_group}{MultiAssayExperiment}(object)

get_metacell(object, ...)

\S4method{get_metacell}{SummarizedExperiment}(object, slot.name = c("metacell", "qMetacell"))

get_cc(object, ...)

\S4method{get_cc}{SummarizedExperiment}(object)

get_proteinID(object, ...)

\S4method{get_proteinID}{SummarizedExperiment}(object)

get_colID(object, ...)

\S4method{get_colID}{SummarizedExperiment}(object)

get_type(object, ...)

\S4method{get_type}{SummarizedExperiment}(object)

get_pkg_version(object, ...)

\S4method{get_pkg_version}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{An instance of class \code{SummarizedExperiment}.}

\item{...}{xxxx description}

\item{slot.name}{The name of the slot dedicated to cell metadata to search.
Default values are 'metacell' and 'qMetacell'}
}
\value{
See individual method description for the return value.

If exists, the slot value requested.

A DataFrame containing the adjacency matrix of the dataset

A data.frame containing the metadata of the dataset

A data.frame containing the metadata of the dataset

A data.frame containing the metadata of the dataset

A data.frame containing the metadata of the dataset

A data.frame containing the metadata of the dataset

A data.frame containing the metadata of the dataset

A data.frame containing the metadata of the dataset
}
\description{
xxx
}
\examples{

## -----------------------------------
## Accessing slots from a MSnSet dataset
## -----------------------------------
data(sub_R25)
se1 <- sub_R25[[1]]
proteinID <- get_proteinID(se1)
colID <- get_colID(se1)
type <- get_type(se1)
metacell <- get_metacell(se1)
conds <- get_group(sub_R25)

}
