% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{actionBtnClass}
\alias{actionBtnClass}
\title{xxx}
\format{
An object of class \code{character} of length 1.
}
\usage{
actionBtnClass
}
\description{
xxx
}
\section{Fields}{

\describe{
\item{\code{actionBtnClass}}{xxx}
}}

\keyword{datasets}
