% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add-module}
\alias{add-module}
\alias{addModule}
\title{Add shiny module}
\usage{
addModule(addons = list())
}
\arguments{
\item{addons}{A \code{list} in which each item is:
\itemize{
\item is named by the name of a package
\item contains th set of modules to integrate
}}
}
\description{
xxxx
}
