% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addImgPath}
\alias{addImgPath}
\title{Add resource path for image}
\usage{
addImgPath(prefix, path)
}
\arguments{
\item{prefix}{A \code{character()} vector which contains packages names}

\item{path}{xxx}
}
\description{
This line allows to extend the resource path to a different directory
than the one in 'inst/' which is given by default when loading a package.
}
\examples{
\donttest{
addImgPath('mod_foo', system.file('.', package='DaparViz'))
}

}
\author{
Samuel Wieczorek
}
