% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addModules}
\alias{addModules}
\title{Add shiny module}
\usage{
addModules(addons)
}
\arguments{
\item{addons}{A \code{list} in which each item is:
\itemize{
\item is named by the name of a package
\item contains th set of modules to integrate
}}
}
\description{
xxxx
}
