% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{bad_format_txt}
\alias{bad_format_txt}
\title{xxx}
\format{
An object of class \code{character} of length 1.
}
\usage{
bad_format_txt
}
\description{
xxx
}
\section{Fields}{

\describe{
\item{\code{bad_format_txt}}{xxx}
}}

\keyword{datasets}
