% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{boxPlot}
\alias{boxPlot}
\title{Builds a boxplot from a dataframe using the package \code{highcharter}}
\usage{
boxPlot(obj, conds, legend = NULL, pal = NULL, subset = NULL)
}
\arguments{
\item{obj}{Numeric matrix}

\item{conds}{xxx}

\item{legend}{A vector of the conditions (one condition per sample).}

\item{pal}{A basis palette for the boxes which length must be equal
to the number of unique conditions in the dataset.}

\item{subset}{A vector of index indicating which rows to highlight}
}
\value{
A boxplot
}
\description{
Builds a boxplot from a dataframe using the package \code{highcharter}
}
\examples{
data(sub_R25)
conds <- legend <- colData(sub_R25)$group
pal <- ExtendPalette(length(unique(conds)))
boxPlot(sub_R25[[1]], conds, legend, pal, seq_len(10))
}
\author{
Samuel Wieczorek, Anais Courtier, Enora Fremy
}
