% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bsmodal.R
\name{bsmodal}
\alias{bsmodal}
\alias{bsmodal_ui}
\alias{bsmodal_server}
\title{Predefined modal}
\usage{
bsmodal_ui(id)

bsmodal_server(id, title = NULL, width = NULL, uiContent = NULL)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the instance of the module}

\item{title}{A \code{character(1)}}

\item{width}{A \code{character(1)} indicating the size of the modal window. Can
be "s" for small (the default), "m" for medium, or "l" for large.}

\item{uiContent}{The content of the modal dialog.}
}
\value{
A Shiny modal-dialog
}
\description{
Displays of formatted modal-dialog with 'Cancel' and
'Ok' buttons.
}
