% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_datasets.R
\name{build-example-datasets}
\alias{build-example-datasets}
\alias{build_toylist_example}
\alias{build_VizData_example}
\title{Build toy datasets for omXplore modules}
\usage{
build_toylist_example(name)

build_VizData_example(name = "obj")
}
\arguments{
\item{name}{The name of the dataset}
}
\description{
xxxxx
}
\examples{
build_toylist_example('myData')

build_VizData_example()


}
