% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Build_enriched_qdata}
\alias{Build_enriched_qdata}
\title{xxxx}
\usage{
Build_enriched_qdata(obj)
}
\arguments{
\item{obj}{An instance of the class \code{SummarizedExperiment}}
}
\value{
A list
}
\description{
xxxx
}
