% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_datasets.R
\name{build_example_datasets}
\alias{build_example_datasets}
\alias{build_toylist_example}
\alias{build_VizData_example}
\alias{build_VizList_example}
\alias{build_VizList_example_from_VizData}
\title{Build toy datasets for omXplore modules}
\usage{
build_toylist_example(name)

build_VizData_example(name = "obj")

build_VizList_example(name = "obj")

build_VizList_example_from_VizData(name = "obj")
}
\arguments{
\item{name}{The name of the dataset}
}
\description{
xxxxx
}
\examples{
build_toylist_example('myData')
build_VizData_example()

}
