% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BuildColorStyles}
\alias{BuildColorStyles}
\title{Build color style for DT tables}
\usage{
BuildColorStyles(type)
}
\arguments{
\item{type}{The type of dataset. Available values are \code{protein} and \code{peptide}}
}
\value{
A list
}
\description{
This function builds a list which is used for styling DT tables with the
function \code{DT::styleEqual()}
}
\examples{
NULL


}
