% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_compare_assays.R
\name{plotCompareAssays}
\alias{plotCompareAssays}
\title{Compare two assays}
\usage{
plotCompareAssays(
  assay1,
  assay2,
  conds,
  info = NULL,
  pal.name = NULL,
  subset.view = NULL,
  n = 100,
  type = "scatter",
  FUN = NULL
)
}
\arguments{
\item{assay1}{A numeric matrix containing quantitative data before
normalization.}

\item{assay2}{A numeric matrix containing quantitative data after
normalization.}

\item{conds}{A vector of the conditions (one condition per sample).}

\item{info}{xxx}

\item{pal.name}{xxx}

\item{subset.view}{xxx}

\item{n}{xxx}

\item{type}{scatter or line}

\item{FUN}{xxx}
}
\value{
A plot
}
\description{
This plot compares the quantitative proteomics data between two assays.
It can be used for example to compare the effect of the normalization
process.

The comparison is made with the division operator.
}
\examples{
data(ft, package='DaparViz')
assay1 <- assay(ft, 1)
assay2 <- 3 * assay(ft, 1)
conds <- colData(ft)$Condition

## -----------------------------------------------------------------
##
## -----------------------------------------------------------------

plotCompareAssays(assay1, assay2, conds, n = 5)

}
\author{
Samuel Wieczorek, Enora Fremy
}
