% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_cc.R
\name{mod_plots_cc_ui}
\alias{mod_plots_cc_ui}
\alias{mod_plots_cc_server}
\title{mod_plots_cc_ui and mod_plots_cc_server}
\usage{
mod_plots_cc_ui(id)

mod_plots_cc_server(input, output, session, cc, matAdj, obj)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{cc}{xxx}

\item{matAdj}{xxx}

\item{obj}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{

xxx

}
\keyword{internal}
