% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizData_Class.R
\name{Convert2VizData,MSnSet-method}
\alias{Convert2VizData,MSnSet-method}
\alias{Convert2VizData,list-method}
\alias{CheckClass}
\alias{convert2viz}
\title{Get the class of object}
\usage{
\S4method{Convert2VizData}{MSnSet}(object, ...)

\S4method{Convert2VizData}{list}(object, ...)

CheckClass(ll)

convert2viz(obj = NULL)
}
\arguments{
\item{object}{xxx}

\item{...}{xxxx}

\item{ll}{An object.}

\item{obj}{An instance of class \code{VizList}}
}
\value{
An instance of class \code{VizList}
}
\description{
This function returns the class of the object passed as argument,
or the class of its items if the object is a \code{list}. If the class is not managed
by DaparViz, then return NA For a list of managed classes.
see \code{FormatAvailables()}.

Actually, three types of dataset can be converted:
\itemize{
\item A \code{list} which must be formatted in the correct format. See xxx
\item A \code{list} of instances of class \code{MSnset}
\item An instance of class \code{QFeatures} (which is already a list)
}
}
