% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_accessors.R
\name{convertMAEtype}
\alias{convertMAEtype}
\title{xxx}
\usage{
convertMAEtype(object)
}
\arguments{
\item{object}{An object that inherits MultiAssayExperiment class
(e.g. MultiAssayExperiment, QFeatures)}
}
\value{
An object of class DaparViz.
}
\description{
xxx
}
\examples{
NULL

}
