% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{create_ft_example}
\alias{create_ft_example}
\title{Creates an example dataset}
\usage{
create_ft_example(with.na = FALSE)
}
\arguments{
\item{with.na}{A \code{logical(1)} which indicates whether the
example dataset contains missing values or not. Default is FALSE.}
}
\value{
An instance of class \code{QFeatures}
}
\description{
This function builds a dataset which is an instance of
the class \code{QFeatures}.
}
\examples{
create_ft_example()
create_ft_example(TRUE)

}
\author{
Samuel Wieczorek
}
