% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{customChart}
\alias{customChart}
\title{Customised resetZoom Button of highcharts plots}
\usage{
customChart(
  hc,
  chartType = "scatter",
  zoomType = "None",
  width = 0,
  height = 0
)
}
\arguments{
\item{hc}{A highcharter object}

\item{chartType}{The type of the plot}

\item{zoomType}{The type of the zoom (one of "x", "y", "xy", "None")}

\item{width}{The width of the plot}

\item{height}{The height of the plot}
}
\value{
A highchart plot
}
\description{
Customised resetZoom Button of highcharts plots
}
\examples{
if (interactive()) {
    library(highcharter)
    hc <- highchart()
    hc_chart(hc, type = "line")
    hc_add_series(hc, data = c(29, 71, 40))
    customChart(hc)
}

NULL

}
\author{
Samuel Wieczorek
}
