% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{customExportMenu}
\alias{customExportMenu}
\title{Customised contextual menu of highcharts plots}
\usage{
customExportMenu(hc, fname)
}
\arguments{
\item{hc}{A highcharter object}

\item{fname}{The filename under which the plot has to be saved}
}
\value{
A contextual menu for highcharts plots
}
\description{
Customised contextual menu of highcharts plots
}
\examples{
NULL

}
\author{
Samuel Wieczorek
}
