% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_funcs.R
\name{DaparViz-converter}
\alias{DaparViz-converter}
\alias{convert2Viz,ANY-method}
\alias{convert2Viz,MultiAssayExperiment-method}
\alias{convert2Viz,QFeatures-method}
\alias{convert2Viz,SummarizedExperiment-method}
\alias{convert2Viz,MSnSet-method}
\title{Convert datasets to \code{DaparViz} class}
\usage{
\S4method{convert2Viz}{ANY}(object)

\S4method{convert2Viz}{MultiAssayExperiment}(object)

\S4method{convert2Viz}{QFeatures}(object)

\S4method{convert2Viz}{SummarizedExperiment}(object)

\S4method{convert2Viz}{MSnSet}(object, ...)
}
\arguments{
\item{object}{An instance of class \code{MSnSet}.}

\item{...}{xxx}
}
\value{
An object of class \code{DaparViz}.
}
\description{
These functions are converters for different data formats
to an object of class \code{DaparViz} which is the format used by functions
in the package \code{DaparViz}.
}
\examples{

# Conversion of an object of class `QFeatures`
data(vData_ft)
convert2Viz(vData_ft)

# Conversion of an object of class `MSnSet`
data(vData_ms)
convert2Viz(vData_ms)

# Conversion of an object of class `SummarizedExperiment`
data(miniACC, package='MultiAssayExperiment')
convert2Viz(miniACC[[1]])


# Conversion of an object of class `MultiAssayExperiment`
data(miniACC, package='MultiAssayExperiment')
convert2Viz(miniACC)

}
