% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparViz_tabExplorer.R
\name{DaparViz_tabExplorer}
\alias{DaparViz_tabExplorer}
\alias{DaparViz_tabExplorer_ui}
\alias{DaparViz_tabExplorer_server}
\title{Explore a list of \code{DaparViz} objects.}
\usage{
DaparViz_tabExplorer_ui(id)

DaparViz_tabExplorer_server(
  id,
  obj = reactive({
     NULL
 }),
  digits = reactive({
     3
 })
)

DaparViz_tabExplorer(obj)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An instance of the class \code{DaparViz}}

\item{digits}{xxx}
}
\value{
NA

NA

NA

A shiny app
}
\description{
xxx
}
\examples{
if (interactive()) {
  data(vData_ft)
  DaparViz_tabExplorer(vData_ft[[1]])
}

}
