% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparVizData.R
\docType{class}
\name{DaparVizData}
\alias{DaparVizData}
\alias{DaparVizData-class}
\alias{initialize,DaparVizData-method}
\alias{Build_DaparVizData,list-method}
\title{DaparVizData class definition}
\usage{
\S4method{initialize}{DaparVizData}(.Object, name, qdata, metacell, metadata, colID, conds, type)

\S4method{Build_DaparVizData}{list}(object)
}
\arguments{
\item{.Object}{xxx}

\item{name}{xxx}

\item{qdata}{xxx}

\item{metacell}{xxx}

\item{conds}{xxx}

\item{type}{xxx}

\item{fullname}{xxx}

\item{mode}{xxx}

\item{steps}{xxx}

\item{mandatory}{xxx

#' @exportMethod qMetacell
#' @rdname QFeatures-accessors
#' @return NA}
}
\description{
This class is used to store the configuration of any process
used with MagellanNTK It contains a validity function to ensure
that the format is correct.

Validity:
\itemize{
\item The first step must be called 'Description', it is a mandatory step. Thus,
the first item of the mandatory vector is TRUE.
To be continued...
}
\subsection{Initialization}{
\subsection{Generic process}{

A generic process
\itemize{
\item Generic pipeline : xxxx
\item Description pipeline: This case is for a process -called 'Description' which is
the first process module of a pipeline
}
}

}

Wrapper function to the constructor of the class
}
\section{Slots}{

\describe{
\item{\code{name}}{xxx}

\item{\code{qdata}}{xxx}

\item{\code{metacell}}{xxx}

\item{\code{conds}}{xxx}

\item{\code{type}}{xxx}
}}

\examples{
library(MSnbase)
library(DaparToolshed)


data(Exp1_R25_prot, package='DaparToolshedData')
qf <- Exp1_R25_prot
vizData <- Build_DaparVizData(qf, 1)



data(Exp1_R25_pept, package='DAPARdata')
msnset <- Exp1_R25_pept
Build_DaparVizData(msnset)

ll <- list(
  name = 'test_list',
  qdata = exprs(msnset),
  metacell = fData(msnset)[ , msnset@experimentData@other$names_metacell],
  mdata = fData(msnset),
  colID = msnset@experimentData@other$keyId,
  conds = pData(msnset)[, 'Condition'],
  type = msnset@experimentData@other$typeOfData
  )

Build_DaparVizData(ll)
}
