% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_cc.R
\name{ds-cc}
\alias{ds-cc}
\alias{omXplore_cc_ui}
\alias{omXplore_cc_server}
\alias{omXplore_cc}
\title{plots_cc_ui and plots_cc_server}
\usage{
omXplore_cc_ui(id)

omXplore_cc_server(
  id,
  obj = reactive({
     NULL
 }),
  i = reactive({
     1
 })
)

omXplore_cc(obj, i)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An instance of \code{SummarizedExperiment} class}

\item{i}{xxx}
}
\value{
A shiny module

A shiny plot

A shiny module

A shiny app

A shiny app
}
\description{
A shiny Module.
}
\examples{
if (!interactive()) {
  data(vdata)
  omXplore_cc(vdata, 1)
}

}
\keyword{internal}
