% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_tabExplorer.R
\name{ds_tabExplorer}
\alias{ds_tabExplorer}
\alias{mod_ds_tabExplorer_ui}
\alias{mod_ds_tabExplorer_server}
\alias{ds_explorer}
\title{Explore a list of \code{DaparViz} objects.}
\usage{
mod_ds_tabExplorer_ui(id)

mod_ds_tabExplorer_server(
  id,
  obj = reactive({
     NULL
 }),
  digits = reactive({
     3
 })
)

ds_explorer(obj)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An instance of the class \code{DaparViz}}

\item{digits}{xxx}
}
\value{
NA
}
\description{
xxx
}
\examples{
if(interactive()){
data(vData_ft)
ds_explorer(vData_ft[[1]])
}

}
