% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_apps_examples.R
\name{external_app}
\alias{external_app}
\alias{extFoo1_ui}
\alias{extFoo1_server}
\alias{extFoo1}
\alias{extFoo2_ui}
\alias{extFoo2_server}
\alias{extFoo2}
\title{External module example}
\usage{
extFoo1_ui(id)

extFoo1_server(
  id,
  dataIn = reactive({
     NULL
 }),
  i = reactive({
     NULL
 })
)

extFoo1(dataIn, i)

extFoo2_ui(id)

extFoo2_server(
  id,
  dataIn = reactive({
     NULL
 }),
  i = reactive({
     NULL
 })
)

extFoo2(dataIn, i)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{dataIn}{An object of instance \code{MultiAssayExperiment}}

\item{i}{An integer which is the index of the assay in the param obj}
}
\value{
NA

NA

NA

A shiny app

NA

NA

A shiny app
}
\description{
Example for an external shiny module, well structured to be run within a
workflow for \code{MagellanNTK}
}
\examples{
if (interactive()) {
    data(vdata)
    app1 <- extFoo1(vdata, 1)
    app2 <- extFoo2(vdata, 1)
    shiny::runApp(app1)
    shiny::runApp(app2)
}

}
