% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizData_accessors.R
\name{ExtractInfos}
\alias{ExtractInfos}
\title{xxx}
\usage{
ExtractInfos(object)
}
\arguments{
\item{object}{An object that inherits \code{MultiAssayExperiment} class
(e.g. MultiAssayExperiment, QFeatures)}
}
\value{
An object of class VizData.
}
\description{
xxx
}
\examples{
NULL

}
