% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparViz_format_DT.R
\name{format_DT}
\alias{format_DT}
\alias{DaparViz_formatDT_ui}
\alias{DaparViz_formatDT_server}
\title{formatDT_ui and formatDT_server}
\usage{
DaparViz_formatDT_ui(id)

DaparViz_formatDT_server(
  id,
  data = reactive({
     NULL
 }),
  data_nostyle = reactive({
     NULL
 }),
  withDLBtns = FALSE,
  showRownames = FALSE,
  dom = "Bt",
  dt_style = reactive({
     NULL
 }),
  filename = "Prostar_export",
  hideCols = reactive({
     NULL
 }),
  selection = "single"
)
}
\arguments{
\item{id}{shiny id}

\item{data}{A \code{data.frame}}

\item{data_nostyle}{xxx}

\item{withDLBtns}{A boolean to indicate whether to display download buttons or not.}

\item{showRownames}{A boolean to indicate whether to show rownames.}

\item{dom}{xxx}

\item{dt_style}{A \code{list} composed of:
\itemize{
\item data : a data.frame` xxxx
\item colors : a named vector
}}

\item{filename}{A \code{character(1)} which is the default filename for download.}

\item{hideCols}{xxx}

\item{selection}{xxx}
}
\value{
NA
}
\description{
A shiny Module.

See \code{DT} package homepage for more details about styling tables.
If no style is precised, this module show the raw data.
If any style is given, then the dataset must be well configured (I.e. it
must contain the correct columns )
}
\examples{
if(interactive()){
data(vData_ft)
DaparViz_formatDT(GetSlotQdata(vData_ft[[1]]))
}


}
