% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{FormatAvailables}
\alias{FormatAvailables}
\title{Formats available in omXplore}
\usage{
FormatAvailables()
}
\value{
A vector of classes compliant with the package omXplore
}
\description{
Gives the list of classes for the datasets which be can converted by the
function \code{convert2VizList()}.
}
\examples{
FormatAvailables()

}
