% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{ft}
\alias{ft}
\title{ft dataset}
\format{
An object of class \code{QFeatures}
}
\usage{
data(ft, package = 'DaparViz')
}
\description{
xxx
}
\details{
ft dataset
}
\keyword{data}
\keyword{datasets}
