% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{general_style}
\alias{general_style}
\title{xxx}
\format{
An object of class \code{character} of length 1.
}
\usage{
general_style
}
\description{
xxx
}
\section{Fields}{

\describe{
\item{\code{general_style}}{xxx}
}}

\keyword{datasets}
