% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetCCInfos}
\alias{GetCCInfos}
\title{xxx}
\usage{
GetCCInfos(cc)
}
\arguments{
\item{cc}{xxx}
}
\value{
A \code{list} of three items:
\itemize{
\item \code{One_One}: the number of cc composed of one protein and one peptide
\item \code{One_Multi}: the number of cc composed of one protein and several peptides
\item \code{Multi_Multi}: the number of cc composed of several proteins and
several (shared) peptides.
}
}
\description{
xxxx
}
\examples{
data(vData_ft)
GetCCInfos(GetSlotCc(vData_ft[[1]]))

}
