% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell_utils.R
\name{GetMetacellTags}
\alias{GetMetacellTags}
\title{Get metacell tags}
\usage{
GetMetacellTags(object = NULL, level = NULL, onlyPresent = FALSE)
}
\arguments{
\item{object}{xxx}

\item{level}{xxx}

\item{onlyPresent}{}
}
\description{
xxx
}
\examples{
NULL

}
\author{
Samuel Wieczorek
}
