% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetPkgVersion}
\alias{GetPkgVersion}
\title{Package version}
\usage{
GetPkgVersion(pkg)
}
\arguments{
\item{pkg}{The name of the package}
}
\value{
A \code{character(1)} with the name of the package and
its version number.
}
\description{
Gets the version number of a package
}
\examples{
GetPkgVersion('omXplore')

}
