% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{heatmapD}
\alias{heatmapD}
\title{This function is a wrapper to \code{heatmap.2} that displays
quantitative data in the \code{Biobase::exprs()} table of an object of
class \code{MSnSet}}
\usage{
heatmapD(vData, distance = "euclidean", cluster = "complete", dendro = FALSE)
}
\arguments{
\item{distance}{The distance used by the clustering algorithm to compute
the dendrogram. See \code{help(heatmap.2)}}

\item{cluster}{the clustering algorithm used to build the dendrogram.
See \code{help(heatmap.2)}}

\item{dendro}{A boolean to indicate fi the dendrogram has to be displayed}

\item{qData}{A dataframe that contains quantitative data.}

\item{conds}{A vector containing the conditions}
}
\value{
A heatmap
}
\description{
This function is a wrapper to \code{heatmap.2} that displays
quantitative data in the \code{Biobase::exprs()} table of an object of
class \code{MSnSet}
}
\examples{

vList <- BuildExampleDataset('MSnbase')
vData <- vList@ll.vizData[[1]]
heatmapD(vData)
mv.heatmap(vData@qdata)

#------------------------------------------
# Shiny module
#------------------------------------------


ui <- mod_ds_heatmap_ui("plot")

server <- function(input, output, session) {
  
  vList <- BuildExampleDataset('MSnbase')
  vData <- vList@ll.vizData[[1]]
  
  mod_ds_heatmap_server("plot",
                        reactive({vData})
  )
  }
shinyApp(ui = ui, server = server)
}
\author{
Florence Combes, Samuel Wieczorek, Enor Fremy
}
