% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.listOf}
\alias{is.listOf}
\title{Checks the class of a list's slots}
\usage{
is.listOf(object, obj.class = NULL)
}
\arguments{
\item{object}{A \code{list}}

\item{obj.class}{The name of the class to search in items of the list.}
}
\value{
A \code{character(1)} with the name of the package or NULL
}
\description{
Checks if all slots of the given list are of the same class.
}
\examples{
ll <- as.list(LETTERS[1:3])
is.listOf(ll, "data.frame")
is.listOf(ll, "character")

}
