% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_metacell.R, R/plot_metacell.R
\name{metacell-plots}
\alias{metacell-plots}
\alias{mod_ds_metacell_ui}
\alias{mod_ds_metacell_server}
\alias{metacellPerLinesHisto_HC}
\alias{metacellPerLinesHistoPerCondition_HC}
\alias{metacellHisto_HC}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_metacell_ui(id)

mod_ds_metacell_server(
  id,
  vizData,
  pal = reactive({
     NULL
 }),
  pattern = reactive({
     NULL
 }),
  showSelect = reactive({
     TRUE
 })
)

metacellPerLinesHisto_HC(
  vizData,
  pattern = NULL,
  detailed = FALSE,
  indLegend = "auto",
  showValues = FALSE
)

metacellPerLinesHistoPerCondition_HC(
  vizData,
  pattern = NULL,
  indLegend = "auto",
  showValues = FALSE,
  pal = NULL
)

metacellHisto_HC(
  vizData,
  pattern = NULL,
  indLegend = "auto",
  showValues = FALSE,
  pal = NULL
)
}
\arguments{
\item{pattern}{xxx}

\item{detailed}{'value' or 'percent'}

\item{indLegend}{The indices of the column name's in \code{Biobase::pData()}
tab}

\item{showValues}{A logical that indicates whether numeric values should be
drawn above the bars.}

\item{qdata}{An instance of the class \code{SummarizedExperiment}}

\item{metacell}{xxx}

\item{type}{xxx}

\item{design}{xxx}
}
\value{
NA

A bar plot
}
\description{
xxxx

This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
}
\examples{
library(MSnbase)
library(DaparToolshed)

#vList <- BuildExampleDataset('QFeatures')
vList <- BuildExampleDataset('MSnbase')
#vList <- BuildExampleDataset('list')

vData <- GetVizData(vList, 1)
pal <- ExtendPalette(length(unique(vData@conds)), "Dark2")


metacellPerLinesHisto_HC(vData, pattern = "Missing POV")
metacellPerLinesHisto_HC(vData)
metacellPerLinesHisto_HC(vData, pattern = "Quantified")
metacellPerLinesHisto_HC(vData, pattern = "Quant. by direct id")
metacellPerLinesHisto_HC(vData, pattern = "Quant. by recovery")
metacellPerLinesHisto_HC(vData, pattern = c("Quantified", "Quant. by direct id", "Quant. by recovery"))


metacellPerLinesHistoPerCondition_HC(vData, pattern = "Missing POV")
metacellPerLinesHistoPerCondition_HC(vData)
metacellPerLinesHistoPerCondition_HC(vData, pattern = "Quantified")
metacellPerLinesHistoPerCondition_HC(vData, pattern = "Quant. by direct id")
metacellPerLinesHistoPerCondition_HC(vData, pattern = "Quant. by recovery")
metacellPerLinesHistoPerCondition_HC(vData, pattern = c("Quantified", "Quant. by direct id", "Quant. by recovery"))


metacellHisto_HC(vData, pattern = "Missing POV")
metacellHisto_HC(vData)
metacellHisto_HC(vData, pattern = "Quantified")
metacellHisto_HC(vData, pattern = "Quant. by direct id")
metacellHisto_HC(vData, pattern = "Quant. by recovery")
metacellHisto_HC(vData, pattern = c("Quantified", "Quant. by direct id", "Quant. by recovery"))
data(ft, package="DaparToolshed")
pattern <- "Missing POV"
pal <- ExtendPalette(2, "Dark2")
vizData <- Convert2VizData(ft, 1)
metacellHisto_HC(vizData, pattern, showValues = TRUE, pal = pal)

}
\author{
Florence Combes, Samuel Wieczorek
}
