% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacellPlots.R
\name{metacellPerLinesHisto_HC}
\alias{metacellPerLinesHisto_HC}
\title{Bar plot of missing values per lines using highcharter}
\usage{
metacellPerLinesHisto_HC(
  object,
  design,
  pattern = NULL,
  detailed = FALSE,
  indLegend = "auto",
  showValues = FALSE
)
}
\arguments{
\item{design}{xxx}

\item{pattern}{xxx}

\item{detailed}{'value' or 'percent'}

\item{indLegend}{The indice of the column name's in \code{Biobase::pData()}
tab}

\item{showValues}{A logical that indicates whether numeric values should be
drawn above the bars.}

\item{obj}{An instance of the class \code{SummarizedExperiment}}
}
\value{
A bar plot
}
\description{
This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
}
\examples{
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept

obj <- obj[1:10]
se <- obj[[1]]
design <- colData(obj)


pal <- ExtendPalette(length(unique(design(obj)$Condition)), "Dark2")


metacellPerLinesHisto_HC(object = se, design = design, pattern = "Missing POV")
metacellPerLinesHisto_HC(object = se, design = design)
metacellPerLinesHisto_HC(object = se, design = design, pattern = "Quantified")
metacellPerLinesHisto_HC(object = se, design = design, pattern = "Quant. by direct id")
metacellPerLinesHisto_HC(object = se, design = design, pattern = "Quant. by recovery")
metacellPerLinesHisto_HC(object = se, design = design, pattern = c("Quantified", "Quant. by direct id", "Quant. by recovery"))


metacellPerLinesHistoPerCondition_HC(object = se, design = design, pattern = "Missing POV")
metacellPerLinesHistoPerCondition_HC(object = se, design = design)
metacellPerLinesHistoPerCondition_HC(object = se, design = design, pattern = "Quantified")
metacellPerLinesHistoPerCondition_HC(object = se, design = design, pattern = "Quant. by direct id")
metacellPerLinesHistoPerCondition_HC(object = se, design = design, pattern = "Quant. by recovery")
metacellPerLinesHistoPerCondition_HC(object = se, design = design, pattern = c("Quantified", "Quant. by direct id", "Quant. by recovery"))


metacellHisto_HC(object = se, design = design, pattern = "Missing POV")
metacellHisto_HC(object = se, design = design)
metacellHisto_HC(object = se, design = design, pattern = "Quantified")
metacellHisto_HC(object = se, design = design, pattern = "Quant. by direct id")
metacellHisto_HC(object = se, design = design, pattern = "Quant. by recovery")
metacellHisto_HC(object = se, design = design, pattern = c("Quantified", "Quant. by direct id", "Quant. by recovery"))
}
\author{
Florence Combes, Samuel Wieczorek
}
