% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dl.R
\name{mod_dl}
\alias{mod_dl}
\alias{dl_ui}
\alias{dl_server}
\title{mod_dl}
\usage{
dl_ui(id)

dl_server(
  id,
  dataIn = reactive({
     NULL
 }),
  extension = "csv",
  widget.type = "Link",
  name = "foo",
  excel.style = NULL
)
}
\arguments{
\item{id}{internal}

\item{dataIn}{internal}

\item{extension}{Available values are \code{csv} (default), \code{RData} and \code{Excel}.}

\item{widget.type}{Available values are \code{Button} and \code{Link} (default).}

\item{name}{internal}

\item{excel.style}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
library(shiny)

ui <- fluidPage(
  dl_ui("dl")
)

server <- function(input, output, session) {
  data(ft, envir = environment())
  
  dl_server("dl",
            dataIn = reactive({ft}),
            extension = c('csv', 'xlsx', 'RData')
  )
}

if(interactive())
  shinyApp(ui = ui, server = server)
}
