% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_format_DT.R
\name{mod_format_DT}
\alias{mod_format_DT}
\alias{mod_format_DT_ui}
\alias{mod_format_DT_server}
\title{mod_format_DT_ui and mod_format_DT_server}
\usage{
mod_format_DT_ui(id)

mod_format_DT_server(
  id,
  df,
  rownames = FALSE,
  dom = "Bt",
  style = reactive({     NULL })
)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{df}{internal}

\item{rownames}{xxxx}

\item{dom}{xxx}

\item{style}{xxx}
}
\description{
A shiny Module.
}
\examples{
if(interactive()){
 data(ft)
 ui <- mod_format_DT_ui('dt')

 server <- function(input, output, session) {
  mod_format_DT_server(id = 'dt',
                       df = reactive({assay(ft,1)})
                       )
  }
 
 shinyApp(ui=ui, server=server)
}


}
