% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_tracking.R
\name{mod_plots_tracking}
\alias{mod_plots_tracking}
\alias{mod_plots_tracking_ui}
\alias{mod_plots_tracking_server}
\title{mod_plots_tracking_ui and mod_plots_tracking_server}
\usage{
mod_plots_tracking_ui(id)

mod_plots_tracking_server(
  id,
  vizData = reactive({
     NULL
 }),
  resetBtn = reactive({
     FALSE
 })
)
}
\arguments{
\item{id}{shiny id}

\item{vizData}{internal}

\item{resetBtn}{A \code{boolean(1)} which indicates whether to show the 'Reset' button
or not.}

\item{params}{A \code{list} of three items to give instructions to the module
in this it is run in slave mode:
\itemize{
\item type: xxxxx
\item names: xxxxx
\item indices: xxxxx
}}
}
\value{
A \code{list} (same structure as the parameter 'params")
}
\description{
This shiny module provides a tool to select
}
\keyword{internal}
