% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_heatmap.R, R/plot_heatmap.R
\name{omXplore_heatmap}
\alias{omXplore_heatmap}
\alias{omXplore_heatmap_ui}
\alias{omXplore_heatmap_server}
\alias{heatmapD}
\alias{mv.heatmap}
\alias{heatmapForMissingValues}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
omXplore_heatmap_ui(id)

omXplore_heatmap_server(
  id,
  dataIn = reactive({
     NULL
 }),
  i = reactive({
     NULL
 })
)

omXplore_heatmap(dataIn, i)

heatmapD(
  qdata,
  conds,
  distance = "euclidean",
  cluster = "complete",
  dendro = FALSE
)

mv.heatmap(
  x,
  col = grDevices::heat.colors(100),
  srtCol = NULL,
  labCol = NULL,
  labRow = NULL,
  key = TRUE,
  key.title = NULL,
  main = NULL,
  ylab = NULL
)

heatmapForMissingValues(
  x,
  col = NULL,
  srtCol = NULL,
  labCol = NULL,
  labRow = NULL,
  key = TRUE,
  key.title = NULL,
  main = NULL,
  ylab = NULL
)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{dataIn}{An instance of a class \code{MultiAssayExperiment}.}

\item{i}{An integer which is the index of the assay in the param obj}

\item{qdata}{Data}

\item{conds}{A vector of conditions}

\item{distance}{Distance to use}

\item{dendro}{If the dendrogram should be shown}

\item{x}{Data}

\item{col}{Colors}

\item{srtCol}{Sorting column}

\item{labCol}{Column labels}

\item{labRow}{Row labels}

\item{key}{Key}

\item{key.title}{Key title}

\item{main}{Main}

\item{ylab}{Y axis label}
}
\value{
NA

NA

A shiny app

A heatmap

A heatmap

A heatmap
}
\description{
This function is a wrapper to \code{heatmap.2()} that displays
assay data in an instance of \code{SummarizedExperiment}. For
more details, see \code{heatmap.2()}.
}
\examples{
if (interactive()) {
    data(vdata)
    omXplore_heatmap(vdata, 1)
}

}
\author{
Florence Combes, Samuel Wieczorek, Enora Fremy
}
