% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_pca.R
\name{pca-plots}
\alias{pca-plots}
\alias{mod_ds_pca_ui}
\title{Bar plot of missing values per lines using highcharter.}
\usage{
mod_ds_pca_ui(id)
}
\arguments{
\item{id}{xxx.}
}
\value{
A plot
}
\description{
This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
}
\examples{

#----------------------------------------------
# Plots a heatmap for generic quantitative data
#----------------------------------------------

heatmapD(assay(ft, 1), conds)


library(QFeatures)
data(ft)


#----------------------------------------
# Launch a single shiny module
# The list of all plots module is avaiable
# with the function [listPlotModules()].
# In the example, replace 'FOO' by the name of the module
# and add the necessaray parameters
#----------------------------------------

if(interactive()){
 data(ft)
 ui <- mod_ds_pca_ui('plot')

 server <- function(input, output, session) {
  conds <- colData(ft)$Condition
 
  mod_ds_pca_server('plot',
                 reactive({assay(ft, 1)}),
                 colData(ft)$Condition
                 )
  }
 
 shinyApp(ui=ui, server=server)
}


}
\author{
Samuel Wieczorek, Enora Fremy
}
