% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgs.require2}
\alias{pkgs.require2}
\title{Loads packages}
\usage{
pkgs.require2(ll.deps)
}
\arguments{
\item{ll.deps}{A \code{character()} vector which contains packages names}
}
\value{
NA
}
\description{
Checks if a package is available to load it
}
\examples{
pkgs.require2("omXplore")

}
\author{
Samuel Wieczorek
}
