% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_mv.R, R/plot_mv.R
\name{plot-mv}
\alias{plot-mv}
\alias{mod_ds_mv_ui}
\alias{mod_ds_mv_server}
\alias{mvPerLinesHisto}
\alias{mvPerLinesHistoPerCondition}
\alias{mvHisto}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_mv_ui(id)

mod_ds_mv_server(id, data, conds, pal.name = reactive({
    NULL
}))

mvPerLinesHisto(data)

mvPerLinesHistoPerCondition(data, conds, pal.name)

mvHisto(data, conds, showValues = FALSE, pal.name = NULL)
}
\arguments{
\item{id}{xxx}

\item{data}{An instance of the class \link{matrix}}

\item{conds}{A \code{character()}}

\item{pal.name}{xxx}
}
\description{
xxxx
}
\examples{
library(QFeatures)
data(ft_na)
mvPerLinesHisto(assay(ft_na, 1))

mvPerLinesHistoPerCondition(assay(ft_na, 1),
                            colData(ft_na)$Condition)
mvHisto(assay(ft_na, 1), colData(ft_na)$Condition)

#------------------------------------------
# Shiny module
#------------------------------------------
if(interactive()){
 data(ft_na)
 ui <- mod_ds_mv_ui('plot')

 server <- function(input, output, session) {
  mod_ds_mv_server('plot', 
                reactive({assay(ft_na, 1)}),
                colData(ft_na)$Condition)
               }
 
 shinyApp(ui=ui, server=server)
}
}
