% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_variance.R
\name{plot-variance}
\alias{plot-variance}
\alias{omXplore_variance_ui}
\alias{omXplore_variance_server}
\alias{CVDist}
\alias{omXplore_variance}
\title{Variance plot}
\usage{
omXplore_variance_ui(id)

omXplore_variance_server(id, dataIn, i, pal.name = NULL)

CVDist(dataIn, conds, pal.name = NULL)

omXplore_variance(dataIn, i)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{dataIn}{An matrix}

\item{i}{An integer which is the index of the assay in the param obj}

\item{pal.name}{A \code{character(1)} which is the name of the palette from
the package \link[RColorBrewer:ColorBrewer]{RColorBrewer::RColorBrewer} from which the colors are taken. Default
value is 'Set1'.}

\item{conds}{A vector indicating the name of each sample.}
}
\value{
NA

NA

A plot

A shiny app
}
\description{
A shiny module which plots the variance of samples
}
\examples{
if (interactive()) {
    data(vdata)
    shiny::runApp(omXplore_variance(vdata, 1))
}

}
