% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_CC.R
\name{plotCCJitter}
\alias{plotCCJitter}
\title{Display a jitter plot for connected components}
\usage{
plotCCJitter(df, clickFunction = NULL)
}
\arguments{
\item{df}{xxxx}

\item{clickFunction}{xxxx}
}
\value{
A plot
}
\description{
Display a jitter plot for connected components
}
\examples{
data(vData_ft)
obj <- vData_ft[[1]]

n.prot <- unlist(lapply(GetSlotCc(obj), function(x) {ncol(x)}))
n.pept <- unlist(lapply(GetSlotCc(obj), function(x) {nrow(x)}))
df <- tibble::tibble(x = jitter(n.pept),
                    y = jitter(n.prot),
                    index = seq_len(length(GetSlotCc(obj))))
plotCCJitter(df)

}
\author{
Thomas Burger, Samuel Wieczorek
}
