% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_CC.R
\name{plotJitter}
\alias{plotJitter}
\title{Jitter plot of CC}
\usage{
plotJitter(list.of.cc = NULL)
}
\arguments{
\item{list.of.cc}{List of cc such as returned by the function get.pep.prot.cc}
}
\value{
A plot
}
\description{
Jitter plot of CC
}
\examples{
\donttest{
data(vData_ms)
plotJitter(vData_ms[1]@cc)
}

}
\author{
Thomas Burger
}
