% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_CC.R
\name{plotJitter_rCharts}
\alias{plotJitter_rCharts}
\title{Display a jitter plot for connected components}
\usage{
plotJitter_rCharts(df, clickFunction = NULL)
}
\arguments{
\item{df}{xxxx}

\item{clickFunction}{xxxx}
}
\value{
A plot
}
\description{
Display a jitter plot for connected components
}
\examples{
\dontrun{
data(vData_ms)
obj <- vData_ms[1]
ll <- obj@cc[1:4]
n.prot <- unlist(lapply(ll, function(x) {length(x$proteins)}))
n.pept <- unlist(lapply(ll, function(x) {length(x$peptides)}))
df <- tibble::tibble(
x = jitter(n.pept),
y = jitter(n.prot),
index = seq_len(length(ll))
)
plotJitter_rCharts(df)
}

}
\author{
Thomas Burger, Samuel Wieczorek
}
