% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plotPCA_Eigen_hc}
\alias{plotPCA_Eigen_hc}
\title{Plots the eigen values of PCA with the highcharts library}
\usage{
plotPCA_Eigen_hc(res.pca)
}
\arguments{
\item{res.pca}{xxx}
}
\value{
A histogram
}
\description{
Plots the eigen values of PCA with the highcharts library
}
\examples{
library(highcharter)
library(DaparViz)

data(ft, package='DaparToolshed')
vList <- convert2viz(ft)
vData <- vList@ll.vizData[[1]]
res.pca <- wrapper_pca(vData, ncp = 6)
plotPCA_Var(res.pca)
plotPCA_Ind(res.pca)
plotPCA_Eigen_hc(res.pca)




ui <- mod_ds_pca_ui("plot")

server <- function(input, output, session) {
  data(ft, package='DaparToolshed')
  vList <- convert2viz(ft)
  vData <- vList@ll.vizData[[1]]
  
  mod_ds_pca_server("plot", vizData = reactive({vData}))
  }

    shinyApp(ui = ui, server = server)
}
\author{
Samuel Wieczorek
}
