% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plotPCA_Var}
\alias{plotPCA_Var}
\title{Plots variables of PCA}
\usage{
plotPCA_Var(res.pca, chosen.axes = c(1, 2))
}
\arguments{
\item{res.pca}{xxx}

\item{chosen.axes}{The dimensions to plot}
}
\value{
A plot
}
\description{
Plots variables of PCA
}
\examples{
library(highcharter)
library(DaparViz)
library(shiny)

data(Exp1_R25_pept, package='DaparToolshedData')
vList <- convert2viz(Exp1_R25_pept)
vData <- vList@ll.vizData[[1]]


res.pca <- wrapper_pca(vData, ncp = 6)
plotPCA_Var(res.pca)
plotPCA_Ind(res.pca)
plotPCA_Eigen_hc(res.pca)




ui <- mod_ds_pca_ui("plot")

server <- function(input, output, session) {
  data(Exp1_R25_pept, package='DaparToolshedData')
  vList <- convert2viz(Exp1_R25_pept)
  vData <- vList@ll.vizData[[1]]
  
  mod_ds_pca_server("plot", vizData = reactive({vData}))
  }

if (interactive())
  shinyApp(ui = ui, server = server)
}
\author{
Samuel Wieczorek
}
