% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prostar_1x.R
\name{Prostar-1x-compatible}
\alias{Prostar-1x-compatible}
\alias{SE_Compatibility_with_Prostar_1.x}
\alias{MAE_Compatibility_with_Prostar_1x}
\title{Convert datasets exported by the package \code{Prostar}}
\usage{
SE_Compatibility_with_Prostar_1.x(obj, se)

MAE_Compatibility_with_Prostar_1x(obj, mae)
}
\arguments{
\item{obj}{An instance of the class \code{MSnSet}}

\item{se}{An instance of the class \code{SummarizedExperiment}}

\item{mae}{An instance of the class \code{MultiAssayExperiment}}
}
\value{
An enriched instance of the class \code{SummarizedExperiment}

An enriched instance of the class \code{MultiAssayExperiment}
}
\description{
Convert datasets exported by the package \code{Prostar}
}
\examples{
data(sub_R25)



}
