% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_seExplorer.R
\name{SE-explorer}
\alias{SE-explorer}
\alias{mod_ds_seExplorer_ui}
\alias{mod_ds_seExplorer_server}
\title{Explore a \code{VizClass} object}
\usage{
mod_ds_seExplorer_ui(id)

mod_ds_seExplorer_server(
  id,
  DaparViz = reactive({
     NULL
 }),
  digits = reactive({
     3
 })
)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{DaparViz}{An instance of the class \code{DaparViz}}

\item{digits}{xxx}
}
\value{
NA
}
\description{
xxx
}
\examples{
library(DaparViz)


ui <- fluidPage(
  mod_ds_seExplorer_ui("plot")
)

server <- function(input, output, session) {
  data(vData_ft)
  obj <- vData_ft[[2]]
  
  mod_ds_seExplorer_server("plot", reactive({obj}))
  
    }

if (interactive())
  shinyApp(ui = ui, server = server)
}
