% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizData_Class.R
\name{[,VizData,ANY,ANY,ANY-method}
\alias{[,VizData,ANY,ANY,ANY-method}
\title{xxx}
\usage{
\S4method{[}{VizData,ANY,ANY,ANY}(x, i, j = NA, k = NA, l = NA, ..., drop)
}
\arguments{
\item{x}{xxx}

\item{i}{xxx}

\item{j}{xxx}

\item{k}{xxx}

\item{l}{xxx}

\item{...}{xxx}

\item{drop}{xxx}
}
\description{
xxx
}
