% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{sub_Exp1_R2_pept_MSnSet}
\alias{sub_Exp1_R2_pept_MSnSet}
\title{sub_Exp1_R2_pept_MSnSet dataset}
\value{
An instance of the class \code{DaparViz}
}
\description{
xxx
}
\examples{
data(sub_Exp1_R2_pept_MSnSet)

}
\keyword{data}
\keyword{datasets}
