% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{sub_Exp1_R2_prot_MSnSet}
\alias{sub_Exp1_R2_prot_MSnSet}
\title{sub_Exp1_R2_prot_MSnSet dataset}
\description{
xxx
}
\details{
sub_Exp1_R2_prot_MSnSet dataset
}
\keyword{data}
\keyword{datasets}
