% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\name{sub_R25}
\alias{sub_R25}
\title{Feature example data}
\format{
An instance of the class \code{MultiAssayExperiment}
}
\source{
\code{sub_R25} was built from the source code available in
\href{https://github.com/edyp-lab/omXplore/blob/main/inst/scripts/build_datasets.R}{\code{inst/scripts/build_datasets.R}}

The \code{DAPARdata} package: \url{https://github.com/edyp-lab/DAPARdata}
}
\value{
An enriched instance of the class \code{MultiAssayExperiment}
}
\description{
\code{sub_R25} is a protein subset of the dataset 'Exp1_R25_pept' in the
package \code{DAPARdata}.
}
