% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\name{sub_R25_pept}
\alias{sub_R25_pept}
\title{Feature example data}
\source{
\code{sub_R25_pept} was built from the source code available in
\href{https://github.com/prostarproteomics/omXplore/blob/main/inst/scripts/build_datasets.R}{\code{inst/scripts/build_datasets.R}}

The \code{DAPARdata} package: \url{https://github.com/prostarproteomics/DAPARdata}
}
\value{
An enriched instance of the class \code{MultiAssayExperiment}
}
\description{
\code{sub_R25_pept} is a protein subset of the dataset 'sub_R25_pept' in the
package 'DAPARdata'.
}
