% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{sub_R25_prot}
\alias{sub_R25_prot}
\title{Feature example data}
\format{
An object of class \code{MultiAssayExperiment} of length 1.
}
\source{
\code{sub_R25_prot} was built from the source code available in
\href{https://github.com/prostarproteomics/omXplore/blob/main/inst/scripts/build_datasets.R}{\code{inst/scripts/build_datasets.R}}

The \code{DAPARdata} package: \url{https://github.com/prostarproteomics/DAPARdata}
}
\usage{
sub_R25_prot
}
\value{
An enriched instance of the class \code{MultiAssayExperiment}
}
\description{
\code{sub_R25_prot} is a protein subset of the dataset 'Exp1_R25_prot' in the
package 'DAPARdata'.
}
\keyword{datasets}
