% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_seTracker.R
\name{tracking}
\alias{tracking}
\alias{mod_seTracker_ui}
\alias{mod_seTracker_server}
\title{Tracking of entities within plots}
\usage{
mod_seTracker_ui(id)

mod_seTracker_server(id, vizData)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{metadata}{A instance of the class \code{SummarizedExperiment}}

\item{colID}{}
}
\value{
NA
}
\description{
This shiny module offers a UI to select a subset of a dataset
and superimpose quantitative values of this selection on the complete plot
Three modes of selection are implemented:
\itemize{
\item 'Protein list': xxx,
\item 'Random': xxx,
\item 'Specific column': xxx
}
}
