% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{vdata}
\alias{vdata}
\title{Feature example data}
\source{
\code{vdata} was built from the source code available in
\href{https://github.com/prostarproteomics/omXplore/blob/main/inst/scripts/build_datasets.R}{\code{inst/scripts/build_datasets.R}}
}
\value{
An enriched instance of the class \code{MultiAssayExperiment}
}
\description{
\code{vdata} is a small object for testing and
demonstration.
}
\keyword{data}
\keyword{datasets}
