% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{vData_ms}
\alias{vData_ms}
\title{vData_ms dataset}
\value{
An instance of the class \code{DaparViz}
}
\description{
xxx
}
\details{
vData_ms dataset
}
\examples{
data(vData_ms)
}
\keyword{data}
\keyword{datasets}
