% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizData_Class.R
\docType{class}
\name{VizData-class}
\alias{VizData-class}
\alias{VizData}
\alias{show,VizData-method}
\alias{initialize,VizData-method}
\alias{convert2Viz,QFeatures-method}
\alias{convert2Viz,MSnSet-method}
\title{VizData class}
\usage{
\S4method{show}{VizData}(object)

\S4method{initialize}{VizData}(
  .Object,
  qdata,
  metacell,
  metadata,
  colID,
  proteinID,
  conds,
  type,
  adjMat,
  cc
)

\S4method{convert2Viz}{QFeatures}(object)

\S4method{convert2Viz}{MSnSet}(object, ...)
}
\arguments{
\item{object}{An instance of class \code{MSnSet}.}

\item{.Object}{xxx}

\item{qdata}{xxx}

\item{metacell}{xxx}

\item{metadata}{xxx}

\item{colID}{xxx}

\item{proteinID}{xxx}

\item{conds}{xxx}

\item{type}{xxx}

\item{adjMat}{xxx}

\item{cc}{xxx}

\item{...}{xxx}
}
\description{
Conceptually, a \code{VizData} object is a standard representation of all
elements from quantitative other structured data used in proteomics, such
as \code{MSnset} or \code{QFeatures}. It allows to use it as a generic converter.

\code{VizData} objects are not usually created by hand but are created from the \code{VizList}
which is the class accessible by the user.

The recommended way to create \code{VizData} objects is the use the
\code{xxxx()} function of the class \code{VizList}

xxx

xxx
}
\section{Slots}{

\describe{
\item{\code{qdata}}{xxx}

\item{\code{metacell}}{xxx}

\item{\code{metadata}}{xxx}

\item{\code{colID}}{xxx}

\item{\code{proteinID}}{xxx}

\item{\code{conds}}{xxx}

\item{\code{type}}{xxx}

\item{\code{adjMat}}{xxx}

\item{\code{cc}}{xxx}
}}

\examples{
NULL

}
