% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert2VizList}
\alias{convert2VizList}
\title{Convert lists of data to \code{VizList} class}
\usage{
convert2VizList(obj)
}
\arguments{
\item{obj}{A list of lists.}
}
\value{
An instance of class \code{VizList}
}
\description{
These functions are converters for different data formats
to an object of class \code{VizList} which is the format used by functions
in the package \code{omXplore}.
}
\examples{
ll.data1 <- build_toylist_example('data1')
.ll.data2 <- build_toylist_example('data2')
data <- convert2VizList(list(ll.data1))
data <- convert2VizList(list(ll.data1, ll.data2))

}
