% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacellPlots.R
\name{wrapper.mvImage}
\alias{wrapper.mvImage}
\title{Heatmap of missing values from a \code{MSnSet} object}
\usage{
wrapper.mvImage(vizData, pattern = "Missing MEC")
}
\arguments{
\item{pattern}{xxx}

\item{object}{An object of class \code{MSnSet}.}
}
\value{
A heatmap
}
\description{
#' Plots a heatmap of the quantitative data. Each column represent one of
the conditions in the object of class \code{MSnSet} and
the color is proportional to the mean of intensity for each line of
the dataset.
The lines have been sorted in order to vizualize easily the different
number of missing values. A white square is plotted for missing values.
}
\examples{
data(ft, package="DaparToolshed")
vizData <- Convert2VizData(ft, 1)
metacell.mask <- match.metacell(GetMetacell(obj), c("Missing POV", "Missing MEC"), level)
indices <- GetIndices_WholeMatrix(metacell.mask, op = ">=", th = 1)
obj <- MetaCellFiltering(obj, indices, cmd = "delete")
wrapper.mvImage(vizData)

}
\author{
Alexia Dorffer
}
