% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_accessors.R
\name{accessors}
\alias{accessors}
\alias{GetSlotMetadata,}
\alias{GetSlotMetacell,}
\alias{GetSlotQdata,}
\alias{GetSlotProteinID,}
\alias{GetSlotColID,}
\alias{GetSlotConds,}
\alias{GetSlotAdjMat,}
\alias{GetSlotCc}
\alias{GetSlotMetadata}
\alias{GetSlotMetacell}
\alias{GetSlotQdata}
\alias{GetSlotProteinID}
\alias{GetSlotColID}
\alias{GetSlotConds}
\alias{GetSlotType}
\alias{GetSlotAdjMat}
\alias{GetSlotMetadata,SummarizedExperiment-method}
\alias{GetSlotMetadata,MSnSet-method}
\alias{GetSlotQdata,ANY-method}
\alias{GetSlotQdata,SummarizedExperiment-method}
\alias{GetSlotQdata,MSnSet-method}
\alias{GetSlotMetacell,SummarizedExperiment-method}
\alias{GetSlotMetacell,MSnSet-method}
\alias{GetSlotColID,SummarizedExperiment-method}
\alias{GetSlotColID,MSnSet-method}
\alias{GetSlotType,SummarizedExperiment-method}
\alias{GetSlotType,MSnSet-method}
\alias{GetSlotProteinID,SummarizedExperiment-method}
\alias{GetSlotProteinID,MSnSet-method}
\alias{GetSlotConds,MultiAssayExperiment-method}
\alias{GetSlotConds,ANY-method}
\alias{GetSlotConds,MSnSet-method}
\title{Accessors functions}
\usage{
GetSlotMetadata(object, ...)

GetSlotMetacell(object, ...)

GetSlotQdata(object, ...)

GetSlotProteinID(object, ...)

GetSlotColID(object, ...)

GetSlotConds(object, ...)

GetSlotType(object, ...)

GetSlotAdjMat(object, ...)

GetSlotCc(object, ...)

\S4method{GetSlotMetadata}{SummarizedExperiment}(object)

\S4method{GetSlotMetadata}{MSnSet}(object)

\S4method{GetSlotQdata}{ANY}(object)

\S4method{GetSlotQdata}{SummarizedExperiment}(object)

\S4method{GetSlotQdata}{MSnSet}(object)

\S4method{GetSlotMetacell}{SummarizedExperiment}(object)

\S4method{GetSlotMetacell}{MSnSet}(object)

\S4method{GetSlotColID}{SummarizedExperiment}(object)

\S4method{GetSlotColID}{MSnSet}(object)

\S4method{GetSlotType}{SummarizedExperiment}(object)

\S4method{GetSlotType}{MSnSet}(object)

\S4method{GetSlotProteinID}{SummarizedExperiment}(object)

\S4method{GetSlotProteinID}{MSnSet}(object)

\S4method{GetSlotConds}{MultiAssayExperiment}(object)

\S4method{GetSlotConds}{ANY}(object)

\S4method{GetSlotConds}{MSnSet}(object)
}
\arguments{
\item{object}{An instance of class \code{MSnSet}.}

\item{...}{xxxx description}
}
\value{
See individual method description for the return value.

A data.frame containing the metadata of the dataset
}
\description{
xxx

xxx

xxx
}
\examples{

## -----------------------------------
## Accessing slots from a MSnSet dataset
## -----------------------------------
data(sub_Exp1_R2_prot_MSnSet)
obj <- sub_Exp1_R2_prot_MSnSet
metadata <- GetSlotMetadata(obj)
qdata <- GetSlotQdata(obj)
metacell <- GetSlotMetacell(obj)
id <- GetSlotColID(obj)
type <- GetSlotType(obj)
proteinID <- GetSlotProteinID(obj)
conds <- GetSlotConds(obj)

}
