% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizData_accessors.R
\name{accessors}
\alias{accessors}
\alias{GetSlotMetadata,}
\alias{GetSlotMetacell,}
\alias{GetSlotQdata,}
\alias{GetSlotProteinID,}
\alias{GetSlotColID,}
\alias{GetSlotConds,}
\alias{GetSlotAdjMat,}
\alias{GetSlotCc}
\alias{GetSlotMetadata}
\alias{GetSlotMetacell}
\alias{GetSlotQdata}
\alias{GetSlotProteinID}
\alias{GetSlotColID}
\alias{GetSlotConds}
\alias{GetSlotType}
\alias{GetSlotAdjMat}
\alias{GetSlotMetadata,VizData-method}
\alias{GetSlotQdata,VizData-method}
\alias{GetSlotMetacell,VizData-method}
\alias{GetSlotColID,VizData-method}
\alias{GetSlotType,VizData-method}
\alias{GetSlotProteinID,VizData-method}
\alias{GetSlotConds,VizData-method}
\alias{GetSlotCc,ANY-method}
\alias{GetSlotCc,VizData-method}
\alias{GetSlotAdjMat,ANY-method}
\alias{GetSlotAdjMat,VizData-method}
\title{Accessors functions for class \code{VizData}}
\usage{
GetSlotMetadata(object, ...)

GetSlotMetacell(object, ...)

GetSlotQdata(object, ...)

GetSlotProteinID(object, ...)

GetSlotColID(object, ...)

GetSlotConds(object, ...)

GetSlotType(object, ...)

GetSlotAdjMat(object, ...)

GetSlotCc(object, ...)

\S4method{GetSlotMetadata}{VizData}(object)

\S4method{GetSlotQdata}{VizData}(object)

\S4method{GetSlotMetacell}{VizData}(object)

\S4method{GetSlotColID}{VizData}(object)

\S4method{GetSlotType}{VizData}(object)

\S4method{GetSlotProteinID}{VizData}(object)

\S4method{GetSlotConds}{VizData}(object)

\S4method{GetSlotCc}{ANY}(object)

\S4method{GetSlotCc}{VizData}(object)

\S4method{GetSlotAdjMat}{ANY}(object)

\S4method{GetSlotAdjMat}{VizData}(object)
}
\arguments{
\item{object}{An instance of class \code{VizData}.}

\item{...}{xxxx description}
}
\value{
See individual method description for the return value.

If exists, the slot value requested.

A data.frame

A data.frame

A data.frame

A character(0)

A character(0)

A character(0)

A character(0)

A character(0)

A character(0)

A character(0)

A character(0)
}
\description{
Functions to access the slots of the class \code{VizData}.
}
\examples{

## -----------------------------------
## Accessing slots from a MSnSet dataset
## -----------------------------------
toydata <- build_VizData_example()
metadata <- GetSlotMetadata(toydata)
qdata <- GetSlotQdata(toydata)
metacell <- GetSlotMetacell(toydata)
id <- GetSlotColID(toydata)
type <- GetSlotType(toydata)
proteinID <- GetSlotProteinID(toydata)
conds <- GetSlotConds(toydata)

}
