% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\alias{Convert2VizList}
\alias{Convert2VizData}
\title{Placeholder for generics functions documentation}
\usage{
Convert2VizList(object, ...)

Convert2VizData(object, ...)
}
\arguments{
\item{object}{xxx}

\item{...}{xxx}
}
\description{
xxx

xxx
}
