% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{AllGenerics}
\alias{AllGenerics}
\alias{GetSlotColID}
\title{Placeholder for generics functions documentation}
\usage{
GetSlotColID(object, ...)
}
\arguments{
\item{object}{xxx}

\item{...}{xxx}
}
\description{
Placeholder for generics functions documentation
}
