% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_CC.R
\name{buildGraph}
\alias{buildGraph}
\title{Display a CC}
\usage{
buildGraph(cc)
}
\arguments{
\item{cc}{A cc (a list)}
}
\value{
A plot
}
\description{
Display a CC
}
\examples{
\dontrun{
data(Exp1_R25_pept, package="DAPARdata")
obj <- Exp1_R25_pept[seq_len(100)]
X <- BuildAdjacencyMatrix(obj, "Protein_group_IDs", FALSE)
ll <- get.pep.prot.cc(X)
g <- buildGraph(ll[[1]], X)
}

}
\author{
Thomas Burger, Samuel Wieczorek
}
