% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_CC.R
\name{buildGraph}
\alias{buildGraph}
\title{Display a CC}
\usage{
buildGraph(cc)
}
\arguments{
\item{cc}{A cc (a list)}
}
\value{
A plot
}
\description{
Display a CC
}
\examples{
\dontrun{
data(vData_ms)
obj <- vData_ms[1]
g <- buildGraph(obj@cc[[1]], obj@adjMat)
}

}
\author{
Thomas Burger, Samuel Wieczorek
}
