% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_colorLegend.R
\name{color-legend}
\alias{color-legend}
\alias{mod_colorLegend_ui}
\alias{mod_colorLegend_server}
\title{Color legend}
\usage{
mod_colorLegend_ui(id)

mod_colorLegend_server(id, text, colors)
}
\arguments{
\item{id}{The 'id' of the shiny module.}

\item{text}{A \code{character()} xxx}

\item{colors}{A \code{HEX()} xxx}
}
\description{
xxxx
}
\examples{
if(interactive()){
 text <- LETTERS[1:5]
 colors <- RColorBrewer::brewer.pal(5, 'Dark2')
 ui <- mod_colorLegend_ui('plot')
 server <- function(input, output, session) {
   mod_colorLegend_server('plot', text, colors)
   }
 
 shinyApp(ui=ui, server=server)
}
}
