% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_colorLegend.R
\name{color-legend}
\alias{color-legend}
\alias{custom_metacell_colors}
\alias{mod_colorLegend_ui}
\alias{mod_colorLegend_server}
\title{Color legend for DaparToolshed}
\usage{
custom_metacell_colors()

mod_colorLegend_ui(id)

mod_colorLegend_server(
  id,
  presentTags = reactive({
     NULL
 }),
  hide.white = TRUE
)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{presentTags}{A vector of \code{character()} which correspond to the tags.}

\item{hide.white}{A \code{boolean()} to indicate whether the white cells must be
hidden or not.}
}
\value{
NA
}
\description{
Shows xxx based on the tags in the package 'DaparToolshed'
}
\examples{
library(DaparViz)


#-------------------------------------------------------------------
ui <- fluidPage(
  tagList(
    mod_colorLegend_ui("plot1"),
    mod_colorLegend_ui("plot2"),
    mod_colorLegend_ui("plot3")
  )
)

server <- function(input, output, session) {
  
  data(vData_ft)
  obj <- vData_ft[[1]]
  
 tags <- GetMetacellTags(obj@metacell, 
                         level = obj@type, 
                         onlyPresent = TRUE)
  
  # Use the default color palette
  mod_colorLegend_server("plot1", tags)

  # Use of a user-defined color palette
  mod_colorLegend_server("plot2", tags)

  # Use of a  palette
  mod_colorLegend_server("plot3", tags)
  }

if (interactive())
  shinyApp(ui, server)
}
