% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_colorLegend.R
\name{color-legend}
\alias{color-legend}
\alias{mod_colorLegend_ui}
\alias{mod_colorLegend_server}
\title{Color legend}
\usage{
mod_colorLegend_ui(id)

mod_colorLegend_server(id, text, pal.name = NULL, colors = NULL)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{text}{A \code{character()} xxx}

\item{pal.name}{A \code{character()} xxxxx}

\item{colors}{A \code{HEX()} containing the color codes to apply to the plot.
The number of colors must be equal to the length of parameter 'text'.}
}
\description{
xxxx
}
\examples{
if(interactive()){
 text <- LETTERS[1:5]
 colors <- RColorBrewer::brewer.pal(5, 'Dark2')
 ui <- mod_colorLegend_ui('plot')
 server <- function(input, output, session) {
   
   # Use the default color palette
   mod_colorLegend_server('plot', text)
   
   # Use of a user-defined color palette
   mod_colorLegend_server('plot', text, colors = colors)
   
   # Use of a  palette
   mod_colorLegend_server('plot', text, pal.name = "Dark3")
   }
 
 shinyApp(ui=ui, server=server)
}
}
