% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_colorLegend.R
\name{color-legend}
\alias{color-legend}
\alias{mod_colorLegend_ui}
\alias{mod_colorLegend_server}
\title{Color legend}
\usage{
mod_colorLegend_ui(id)

mod_colorLegend_server(id, text, pal.name = NULL, colors = NULL)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{text}{A \code{character()} xxx}

\item{pal.name}{A \code{character()} xxxxx}

\item{colors}{A \code{HEX()} containing the color codes to apply to the plot.
The number of colors must be equal to the length of parameter 'text'.}
}
\value{
NA
}
\description{
xxxx
}
\examples{
if (interactive()) {
    ui <- tagList(
mod_colorLegend_ui("plot1"),
mod_colorLegend_ui("plot2"),
mod_colorLegend_ui("plot3")
)
server <- function(input, output, session) {
   # Use the default color palette
   mod_colorLegend_server("plot1", paste0('Default_',1:5))
   
   # Use of a user-defined color palette
   mod_colorLegend_server("plot2", 
       paste0('Pastel2_',1:5), 
       colors = RColorBrewer::brewer.pal(5, "Pastel2"))
   
   # Use of a  palette
   mod_colorLegend_server("plot3", 
       paste0('Accent_',1:5), 
       pal.name = "Accent")
}

shinyApp(ui, server)
}
}
