% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_colorLegend.R
\name{color-legend}
\alias{color-legend}
\alias{mod_colorLegend_ui}
\alias{mod_colorLegend_server}
\title{Color legend}
\usage{
mod_colorLegend_ui(id)

mod_colorLegend_server(id, presentTags = NULL, hide.white = TRUE)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{presentTags}{A vector of \code{character()} which correspond to the tags.}

\item{hide.white}{xxxxx}
}
\value{
NA
}
\description{
xxxx
}
\examples{
#-------------------------------------------------------------------
ui <- tagList(
  mod_colorLegend_ui("plot1"),
  mod_colorLegend_ui("plot2"),
  mod_colorLegend_ui("plot3")
  )

server <- function(input, output, session) {
  data(Exp1_R25_prot, package='DaparToolshedData')
  obj <- Exp1_R25_prot[[1]]
  tags <- GetMetacellTags(obj, level = typeDataset(obj), onlyPresent = TRUE)
      
  # Use the default color palette
  mod_colorLegend_server("plot1",tags)

  # Use of a user-defined color palette
  mod_colorLegend_server("plot2", tags)

  # Use of a  palette
  mod_colorLegend_server("plot3", tags)
  }

shinyApp(ui, server)
}
