% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectedComps.R, R/mod_explore_graphs.R,
%   R/mod_ds_cc.R
\name{connected-components}
\alias{connected-components}
\alias{wrapper.compute.cc}
\alias{compute.cc}
\alias{convertCC2graph}
\alias{show.graph}
\alias{plotJitter_hc}
\alias{Extract_CC}
\alias{mod_explore_graphs_ui}
\alias{mod_explore_graphs_server}
\alias{mod_plots_cc_ui}
\alias{mod_plots_cc_server}
\title{Connected components of a graph.}
\usage{
wrapper.compute.cc(obj.se)

compute.cc(X)

convertCC2graph(cc, X)

show.graph(g, obj = NULL, prot.tooltip = NULL, pept.tooltip = NULL)

plotJitter_hc(ll.cc, df.tooltips = NULL, clickFunction = NULL)

Extract_CC(ll.cc, type)

mod_explore_graphs_ui(id)

mod_explore_graphs_server(id, ll.cc, se, settings)

mod_plots_cc_ui(id)

mod_plots_cc_server(input, output, session, cc, matAdj, obj, settings)
}
\arguments{
\item{obj.se}{An instance of the class \code{SummarizedExperiment}. This object
must have a adjacencyMatrix.}

\item{X}{A \code{Matrix} which is the adjacency matrix.}

\item{cc}{xxx}

\item{g}{xxx}

\item{obj}{xxx}

\item{prot.tooltip}{A \code{character()} which are the tooltips of the proteins.}

\item{pept.tooltip}{A \code{character()} which are the tooltips of the peptides.}

\item{ll.cc}{xxx}

\item{clickFunction}{A JavaScript code to show an information when a point is clicked}

\item{type}{A \code{character(1)} which denotes the type ofconnected components
to extract. Available values are: 'OneOne', 'OneMulti' and 'MultiAny'.}

\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{se}{An object of class \code{QFeatures}}

\item{settings}{xxx}

\item{matAdj}{xxx}

\item{tooltips}{A \code{data.frame} which contains a subset of the rowData
of the object and will be use as tooltips. Each column of the data.frame
will be showed in the global tooltip of a point in the plot.}
}
\value{
NA
}
\description{
This function computes the connex components of a graph,
represented by its adjacency matrix.

A shiny Module.
}
\examples{
data(ft)
X <- adjacencyMatrix(ft[[1]])

#------------------------------------------
# Build the cc list for an adjaceny matrix
#------------------------------------------

ll.cc <- compute.cc(X)

#-------------------------------------------------
# Convert one cc of this list to a graph structure
#-------------------------------------------------

g <- convertCC2graph(ll.cc[[1]], X)

#-------------------------------------------
# Build the cc list for all type of peptides 
# 'all', 'shared', 'specific'
#-------------------------------------------

ll.cc <- wrapper.compute.cc(ft[[1]])

#-------------------------------------------------
# Show the graph
#-------------------------------------------------
 
show.graph(g)


#-----------------------------------------------------
# Plots a graph which x-axis is the number of peptides
# in the cc and the y-axis is the number of proteins
# in the cc.
#'-----------------------------------------------------

tt <- as.data.frame(rowData(ft[[1]])[,1])
colnames(tt) <- colnames(rowData(ft[[1]]))[1]
plotJitter_hc(ll.cc, tt)
plotJitter_hc(ll.cc)


xxx

}
\seealso{
The \code{\link[=adjacencyMatrix]{adjacencyMatrix()}} function of the package \link{QFeatures}.
}
\author{
Thomas Burger, Samuel Wieczorek
}
\keyword{internal}
