% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_cc.R
\name{connected-components}
\alias{connected-components}
\title{mod_plots_cc_ui and mod_plots_cc_server}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{cc}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{

xxx

}
\keyword{internal}
