% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_pept_prot.R
\name{connected_components}
\alias{connected_components}
\alias{mod_graph_pept_prot_ui}
\alias{mod_graph_pept_prot_server}
\title{Connected components tools.}
\usage{
mod_graph_pept_prot_ui(id)

mod_graph_pept_prot_server(id, cc, matAdj, dataIn, settings)
}
\arguments{
\item{id}{xxx}

\item{cc}{xxx}

\item{matAdj}{xxx}

\item{dataIn}{An object of class \code{QFeatures}}

\item{settings}{xxx}
}
\description{
A shiny Module.
}
