% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_cc.R
\name{connected_components}
\alias{connected_components}
\alias{mod_ds_cc_ui}
\alias{mod_ds_cc_server}
\title{mod_plots_cc_ui and mod_plots_cc_server}
\usage{
mod_ds_cc_ui(id)

mod_ds_cc_server(id, object)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{DaparViz}{An instance of \code{DaparViz} class}
}
\value{
NA
}
\description{
A shiny Module.
}
\keyword{internal}
