% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizData_Class.R
\name{Convert2VizList,QFeatures-method}
\alias{Convert2VizList,QFeatures-method}
\alias{Convert2VizData,MSnSet-method}
\alias{Convert2VizData,list-method}
\alias{CheckClass}
\alias{convert2viz}
\title{xxx}
\usage{
\S4method{Convert2VizList}{QFeatures}(object)

\S4method{Convert2VizData}{MSnSet}(object, ...)

\S4method{Convert2VizData}{list}(object, ...)

CheckClass(ll)

convert2viz(obj)
}
\arguments{
\item{object}{xxx}

\item{...}{xxxx}

\item{ll}{xxx}

\item{obj}{xxx}
}
\value{
An instance of class \link{VizList}
}
\description{
xxx

Actually, three types of dataset can be converted:
\itemize{
\item A \code{list} which must be formatted in the correct format. See xxx
\item A list of instances of class \code{MSnset}
\item An instance of class \code{QFeatures} (which is already a list)
}
}
