% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_mae.R
\name{converters}
\alias{converters}
\alias{convert_to_mae}
\alias{MSnSet_to_mae}
\alias{QFeatures_to_mae}
\alias{SE_to_mae}
\alias{MAE_to_mae}
\alias{Check_se_Consistency}
\alias{listOfSE_to_mae}
\alias{Check_MSnSet_Consistency}
\alias{listOfMSnSet_to_mae}
\title{xxx}
\usage{
convert_to_mae(obj)

MSnSet_to_mae(obj)

QFeatures_to_mae(obj)

SE_to_mae(obj)

MAE_to_mae(obj)

Check_se_Consistency(obj)

listOfSE_to_mae(obj)

Check_MSnSet_Consistency(obj)

listOfMSnSet_to_mae(obj)
}
\arguments{
\item{obj}{An object compliant with formats xxxx}
}
\description{
The resulting object is an instance of the  MultiAssayExperiment class.
F
}
\examples{

#-------------------------------------------
# Conversion of a list of MSnSet instances
#-------------------------------------------
ll.msnset <- list(data1 = Exp1_R25_prot, data2 = Exp1_R25_prot, data3 = Exp1_R25_prot)
convert_to_mae(ll.msnset)

#-------------------------------------------
# Conversion of a MSnSet instance
#-------------------------------------------
convert_to_mae(Exp1_R25_prot)

#-------------------------------------------
# Conversion of a list of SummarizedExperiment instances
#-------------------------------------------



#-------------------------------------------
# Conversion of a SummarizedExperiment instance
#-------------------------------------------
mae <- convert_to_mae(Exp1_R25_prot)
convert_to_mae(mae[[1]])


#-------------------------------------------
# Conversion of a MultiAssayExperiment instance
#-------------------------------------------
mae <- convert_to_mae(Exp1_R25_prot)
convert_to_mae(mae)

}
