% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_corrmatrix.R
\name{corrmatrix}
\alias{corrmatrix}
\alias{omXplore_corrmatrix_ui}
\alias{omXplore_corrmatrix_server}
\alias{corrMatrix}
\alias{omXplore_corrmatrix}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
omXplore_corrmatrix_ui(id)

omXplore_corrmatrix_server(
  id,
  dataIn = reactive({
     NULL
 }),
  i = reactive({
     1
 })
)

corrMatrix(data, rate = 0.5, showValues = FALSE)

omXplore_corrmatrix(dataIn, i)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{dataIn}{An instance of the class \code{SummarizedExperiment}}

\item{i}{An integer which is the index of the assay in the param dataIn}

\item{data}{An object of class 'matrix'}

\item{rate}{The rate parameter to control the exponential law for
the gradient of colors}

\item{showValues}{A boolean which indicates whether to show values in the
correlation plot.}
}
\value{
NA

NA

A plot

A shiny app
}
\description{
Displays a correlation matrix of the quantitative data of a
numeric matrix.
}
\examples{
if (interactive()) {
    data(vdata)
    omXplore_corrmatrix(vdata, 1)
}

}
