% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_corrmatrix.R
\name{corrmatrix}
\alias{corrmatrix}
\alias{omXplore_corrmatrix_ui}
\alias{omXplore_corrmatrix_server}
\alias{corrMatrix}
\alias{omXplore_corrmatrix}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
omXplore_corrmatrix_ui(id)

omXplore_corrmatrix_server(
  id,
  obj = reactive({
     NULL
 }),
  i = reactive({
     1
 })
)

corrMatrix(data, rate = 0.5, showValues = FALSE)

omXplore_corrmatrix(obj, i)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An instance of the class \code{SummarizedExperiment}}

\item{i}{xxx}

\item{data}{An object of class 'matrix'}

\item{rate}{The rate parameter to control the exponential law for
the gradient of colors}

\item{showValues}{xxx}
}
\value{
NA

NA

A plot

A shiny app
}
\description{
xxxx
}
\examples{
if (interactive()) {
  data(vdata)
  omXplore_corrmatrix(vdata, 1)
}

data(vdata)
qdata <- GetSlotQdata(vdata[[1]])
corrMatrix(qdata)

}
