% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_corrmatrix.R, R/plot_corrmatrix.R
\name{corrmatrix}
\alias{corrmatrix}
\alias{mod_ds_corrmatrix_ui}
\alias{mod_ds_corrmatrix_server}
\alias{corrMatrix}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_corrmatrix_ui(id)

mod_ds_corrmatrix_server(
  id,
  data,
  rate = reactive({
     0.5
 }),
  showValues = reactive({
     FALSE
 })
)

corrMatrix(data, rate = 0.5, showValues = FALSE)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{data}{An object of class 'SummarizedExperiment'}

\item{rate}{The rate parameter to control the exponential law for
the gradient of colors}

\item{showValues}{xxx}
}
\value{
NA

NA
}
\description{
xxxx
}
\examples{
library(SummarizedExperiment)
data(ft, package='DaparViz')
corrMatrix(assay(ft, 1))


#------------------------------------------
# Shiny module
#------------------------------------------
if (interactive()) {
    data(ft, package='DaparViz')
    ui <- mod_ds_corrmatrix_ui("plot")

    server <- function(input, output, session) {
        mod_ds_corrmatrix_server(
            "plot",
            reactive({assay(ft, 1)})
        )
    }

    shinyApp(ui = ui, server = server)
}

}
