% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparViz_corrmatrix.R, R/plot_corrmatrix.R
\name{corrmatrix}
\alias{corrmatrix}
\alias{DaparViz_corrmatrix_ui}
\alias{DaparViz_corrmatrix_server}
\alias{DaparViz_corrmatrix}
\alias{corrMatrix}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
DaparViz_corrmatrix_ui(id)

DaparViz_corrmatrix_server(
  id,
  obj = reactive({
     NULL
 }),
  rate = reactive({
     0.5
 }),
  showValues = reactive({
     FALSE
 })
)

DaparViz_corrmatrix(obj)

corrMatrix(data, rate = 0.5, showValues = FALSE)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An instance of the class \code{DaparViz}}

\item{rate}{The rate parameter to control the exponential law for
the gradient of colors}

\item{showValues}{xxx}

\item{data}{An object of class 'matrix'}
}
\value{
NA

NA
}
\description{
xxxx
}
\examples{
if(interactive()){
data(vData_ft)
DaparViz_corrmatrix(vData_ft[[1]])
}


NULL

}
