% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_funcs.R, R/DaparViz-class.R
\docType{class}
\name{convert2Viz,ANY-method}
\alias{convert2Viz,ANY-method}
\alias{convert2Viz,MultiAssayExperiment-method}
\alias{convert2Viz,QFeatures-method}
\alias{convert2Viz,SummarizedExperiment-method}
\alias{convert2Viz,MSnSet-method}
\alias{DaparViz-class}
\alias{DaparViz}
\alias{show,DaparViz-method}
\alias{initialize,DaparViz-method}
\title{DaparViz class}
\usage{
\S4method{convert2Viz}{ANY}(object)

\S4method{convert2Viz}{MultiAssayExperiment}(object)

\S4method{convert2Viz}{QFeatures}(object)

\S4method{convert2Viz}{SummarizedExperiment}(object)

\S4method{convert2Viz}{MSnSet}(object, ...)

\S4method{show}{DaparViz}(object)

\S4method{initialize}{DaparViz}(
  .Object,
  qdata,
  metacell,
  metadata,
  colID,
  proteinID,
  conds,
  type,
  adjMat,
  cc
)
}
\arguments{
\item{object}{An instance of the class \code{DaparViz}}

\item{...}{xxx}

\item{.Object}{xxx}

\item{qdata}{xxx}

\item{metacell}{xxx}

\item{metadata}{xxx}

\item{colID}{xxx}

\item{proteinID}{xxx}

\item{conds}{xxx}

\item{type}{xxx}

\item{adjMat}{xxx}

\item{cc}{xxx}
}
\description{
Conceptually, a \code{DaparViz} object is a standard representation of all
elements from quantitative other structured data used in proteomics, such
as \code{MSnset}, \code{QFeatures}, \code{SummarizedExperiment} or \code{MultAssayExperiment}.
It allows to use it as a generic converter.

\code{DaparViz} objects are not usually created by hand but are created from
the \code{VizList} which is the class accessible by the user.

The recommended way to create \code{DaparViz} objects is the use the
function \code{convert2viz()} of the class \code{VizList}

xxx

xxx
}
\section{Slots}{

\describe{
\item{\code{qdata}}{A matrix containing the quantitative data}

\item{\code{metacell}}{A matrix of xxx}

\item{\code{metadata}}{A matrix of xxxx}

\item{\code{colID}}{The name of the colum which contains the ids of each row.}

\item{\code{proteinID}}{The name of the column in the metadata which represent the
(unique) id of the protein}

\item{\code{conds}}{A vector of the names of the condition of each sample
of the dataset}

\item{\code{type}}{The type of entity in the dataset (peptide, protein, ...)}

\item{\code{adjMat}}{xxx}

\item{\code{cc}}{A list of xxx}
}}

\examples{
NULL

}
