% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{DaparViz-converter}
\alias{DaparViz-converter}
\alias{convert2VizList}
\title{Convert datasets to \code{DaparViz} class}
\usage{
convert2VizList(obj)
}
\arguments{
\item{object}{An instance of class \code{QFeatures}.}
}
\value{
An instance of class \code{VizList}
}
\description{
These functions are converters for different data formats
to an object of class \code{DaparViz} which is the format used by functions
in the package \code{DaparViz}.
}
\examples{

# Conversion of an object of class `MSnSet`
data(sub_Exp1_R2_pept_MSnSet)
obj <- convert2VizList(sub_Exp1_R2_pept_MSnSet)
obj

# Conversion of an object of class `QFeatures`
data(sub_Exp1_R2_pept_qf)
obj <- convert2VizList(sub_Exp1_R2_pept_qf)
obj

}
