% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{DaparViz-modules}
\alias{DaparViz-modules}
\alias{addModules}
\alias{listShinyApps}
\alias{listPlotModules}
\title{Modules functions}
\usage{
addModules(addons)

listShinyApps(location = "both")

listPlotModules(location = "both")
}
\arguments{
\item{addons}{A \code{list} in which each item is:
\itemize{
\item is named by the name of a package
\item contains th set of modules to integrate
}}

\item{location}{xxx}
}
\description{
xxxxx
\itemize{
\item \code{listShinyApps()}: xxx
\item \code{addImgPath()}: xxx
\item \code{addModules()}: xxx
}

xxxx
}
