% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_density.R
\name{density-plot}
\alias{density-plot}
\alias{omXplore_density_ui}
\alias{omXplore_density_server}
\alias{densityPlot}
\alias{omXplore_density}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
omXplore_density_ui(id)

omXplore_density_server(
  id,
  obj = reactive({
     NULL
 }),
  i = reactive({
     NULL
 }),
  pal.name = reactive({
     NULL
 })
)

densityPlot(data, conds = NULL, pal.name = NULL)

omXplore_density(obj, i)
}
\arguments{
\item{id}{xxx}

\item{obj}{An instance of the class \code{SummarizedExperiment}}

\item{i}{xxx}

\item{pal.name}{A \code{character(1)} which is the name of the palette from
the package \link{RColorBrewer} from which the colors are taken. Default
value is 'Set1'.}

\item{data}{xxx}

\item{conds}{xxx}
}
\value{
NA

NA

A plot

A shiny app
}
\description{
xxxx
}
\examples{
if (!interactive()) {
  data(vdata)
  omXplore_density(vdata, 1)
}

data(vdata)
qdata <- SummarizedExperiment::assay(vdata[[1]])
conds <- get_group(vdata)
densityPlot(qdata, conds)

}
