% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_density.R, R/omXplore_format_DT.R
\name{density-plot}
\alias{density-plot}
\alias{omXplore_density_ui}
\alias{omXplore_density_server}
\alias{densityPlot}
\alias{omXplore_density}
\alias{formatDT}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
omXplore_density_ui(id)

omXplore_density_server(
  id,
  obj = reactive({
     NULL
 }),
  pal.name = reactive({
     NULL
 })
)

densityPlot(data, conds, pal.name = NULL)

omXplore_density(obj)

formatDT(obj)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An instance of the class \code{VizData}.}

\item{pal.name}{A \code{character(1)} which is the name of the palette from
the package \link{RColorBrewer} from which the colors are taken. Default
value is 'Set1'.}

\item{data}{A \code{data.frame} or \code{matrix} of \code{numeric()} which is the
quantitative data to plot.}

\item{conds}{A \code{character()} of condition name for each sample. The
length of 'conds' must be equal to the number of columns of 'data'.}
}
\value{
NA

NA

A plot

A shiny app

NA
}
\description{
xxxx
}
\examples{
if (interactive()) {
  data(vData_ft)
  omXplore_density(vData_ft[[1]])
}

data(vData_ft)
qdata <- GetSlotQdata(vData_ft[[1]])
conds <- GetSlotConds(vData_ft[[1]])
densityPlot(qdata, conds)

}
