% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_density.R, R/plot_density.R
\name{density-plot}
\alias{density-plot}
\alias{mod_ds_density_ui}
\alias{mod_ds_density_server}
\alias{densityPlot}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_density_ui(id)

mod_ds_density_server(id, data, conds, pal.name = reactive({
    NULL
}))

densityPlot(data, conds, pal.name = NULL)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{data}{numeric matrix}

\item{conds}{A \code{character()} of condition name for each sample. The
length of 'conds' must be equal to the number of columns of 'qData'.}

\item{pal.name}{A \code{character(1)} which is the name of a palette in
the package \link{RColorBrewer}.}
}
\description{
xxxx
}
\examples{
data(ft)
densityPlot(assay(ft, 1))


#------------------------------------------
# Shiny module
#------------------------------------------
if(interactive()){
data(ft)
 ui <- mod_ds_density_ui('plot')

 server <- function(input, output, session) {
  mod_ds_density_server('plot', 
                     reactive({assay(ft, 1)}),
                     colData(ft)$Condition)
  }
 
 shinyApp(ui=ui, server=server)
}




}
