% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_density.R, R/plot_density.R
\name{density-plot}
\alias{density-plot}
\alias{mod_ds_density_ui}
\alias{mod_ds_density_server}
\alias{densityPlot}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_density_ui(id)

mod_ds_density_server(
  id,
  vizData = reactive({
     NULL
 }),
  pal.name = reactive({
     NULL
 })
)

densityPlot(data, conds, pal.name = NULL)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{pal.name}{A \code{character(1)} which is the name of the palette from
the package \link{RColorBrewer} from which the colors are taken. Default
value is 'Set1'.}

\item{data}{A \code{data.frame} or \code{matrix} of \code{numeric()} which is the
quantitative data to plot.}

\item{conds}{A \code{character()} of condition name for each sample. The
length of 'conds' must be equal to the number of columns of 'data'.}
}
\value{
NA

NA
}
\description{
xxxx
}
\examples{

library(SummarizedExperiment)
data(ft, package='DaparViz')

vizData <- Coerce2VizData(ft,1)

densityPlot(vizData@qdata, conds = vizData@conds)


#------------------------------------------
# Shiny module
#------------------------------------------

data(ft, package='DaparViz')
ui <- shinyUI(mod_ds_density_ui("plot"))

server <- shinyServer(function(input, output, session) {
  mod_ds_density_server("plot", vizData = reactive({vizData})
  )
                         })

shinyApp(ui, server)
}
