% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparViz_density.R, R/plot_density.R
\name{density-plot}
\alias{density-plot}
\alias{DaparViz_density_ui}
\alias{DaparViz_density_server}
\alias{DaparViz_density}
\alias{densityPlot}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
DaparViz_density_ui(id)

DaparViz_density_server(
  id,
  obj = reactive({
     NULL
 }),
  pal.name = reactive({
     NULL
 })
)

DaparViz_density(obj)

densityPlot(data, conds, pal.name = NULL)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An instance of the class \code{DaparViz}.}

\item{pal.name}{A \code{character(1)} which is the name of the palette from
the package \link{RColorBrewer} from which the colors are taken. Default
value is 'Set1'.}

\item{data}{A \code{data.frame} or \code{matrix} of \code{numeric()} which is the
quantitative data to plot.}

\item{conds}{A \code{character()} of condition name for each sample. The
length of 'conds' must be equal to the number of columns of 'data'.}
}
\value{
NA

NA
}
\description{
xxxx
}
\examples{
if(interactive()){
data(vData_ft)
DaparViz_density(vData_ft[[1]])
}

}
