% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_CC.R
\name{display.CC.visNet}
\alias{display.CC.visNet}
\title{Display a CC}
\usage{
display.CC.visNet(
  g,
  layout = layout_nicely,
  obj = NULL,
  prot.tooltip = NULL,
  pept.tooltip = NULL
)
}
\arguments{
\item{g}{A cc (a list)}

\item{layout}{xxxxx}

\item{obj}{xxx}

\item{prot.tooltip}{xxx}

\item{pept.tooltip}{xxx}
}
\value{
A plot
}
\description{
Display a CC
}
\examples{
\donttest{
data(vData_ms)
obj <- vData_ms[1]
display.CC.visNet(obj@cc[[1]], obj@adjMat)
}

}
\author{
Thomas Burger, Samuel Wieczorek
}
