% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_cc.R
\name{ds-cc}
\alias{ds-cc}
\alias{omXplore_cc_ui}
\alias{omXplore_cc_server}
\alias{omXplore_cc}
\title{plots_cc_ui and plots_cc_server}
\usage{
omXplore_cc_ui(id)

omXplore_cc_server(id, obj)

omXplore_cc(obj)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An instance of \code{VizData} class}
}
\value{
A shiny plot

A shiny app

A shiny app
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
  data(vData_ft)
  omXplore_cc(vData_ft[[1]])
}

}
\keyword{internal}
