% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparViz_cc.R
\name{ds-cc}
\alias{ds-cc}
\alias{DaparViz_cc_ui}
\alias{DaparViz_cc_server}
\alias{DaparViz_cc}
\title{plots_cc_ui and plots_cc_server}
\usage{
DaparViz_cc_ui(id)

DaparViz_cc_server(id, obj)

DaparViz_cc(obj)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An instance of \code{DaparViz} class}
}
\value{
A shiny plot

A shiny app

A shiny app
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
  data(vData_ft)
  DaparViz_cc(vData_ft[[1]])
}

}
\keyword{internal}
