% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_pca.R, R/plot_pca.R
\name{ds-pca}
\alias{ds-pca}
\alias{omXplore_pca_ui}
\alias{omXplore_pca_server}
\alias{omXplore_pca}
\alias{wrapper_pca}
\alias{plotPCA_Eigen}
\alias{plotPCA_Var}
\alias{plotPCA_Ind}
\alias{plotPCA_Eigen_hc}
\title{Bar plot of missing values per lines using highcharter.}
\usage{
omXplore_pca_ui(id)

omXplore_pca_server(id, obj, i)

omXplore_pca(obj, i)

wrapper_pca(qdata, group, var.scaling = TRUE, ncp = NULL)

plotPCA_Eigen(res.pca)

plotPCA_Var(res.pca, chosen.axes = c(1, 2))

plotPCA_Ind(res.pca, chosen.axes = c(1, 2))

plotPCA_Eigen_hc(res.pca)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An instance of the class \code{MultiAssayExperiment}.}

\item{i}{An integer which is the index of the assay in the param obj}

\item{qdata}{A data.frame() of quantitative data}

\item{group}{A vector with the name of samples}

\item{var.scaling}{A boolean indicating whether to scale the data or not}

\item{ncp}{See \code{FactoMineR::PCA()}}

\item{res.pca}{The result of the function \code{FactoMineR::PCA()}}

\item{chosen.axes}{See the parameter 'axes' of the function
\code{factoextra::fviz_pca_var()}}
}
\value{
NA

NA

A shiny app

The result of the function FactoMineR::PCA()

A plot

A plot

A plot

A plot
}
\description{
This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
\itemize{
\item \code{wrapper_pca()}
\item \code{plotPCA_Eigen_hc()}: plots the eigen values of PCA with the highcharts
library
\item \code{plotPCA_Eigen()}: plots the eigen values of PCA
\item \code{plotPCA_Var()}
\item \code{plotPCA_Ind()}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
  data(vdata)
  # Replace missing values for the example
  sel <- is.na(SummarizedExperiment::assay(vdata, 1))
  SummarizedExperiment::assay(vdata[[1]])[sel] <- 0
  omXplore_pca(vdata, 1)
\dontshow{\}) # examplesIf}
data(vdata)
obj <- vdata[[1]]
res.pca <- wrapper_pca(SummarizedExperiment::assay(obj), get_group(obj))
plotPCA_Eigen(res.pca)
plotPCA_Var(res.pca)
plotPCA_Eigen_hc(res.pca)
plotPCA_Ind(res.pca)

}
\author{
Samuel Wieczorek, Enora Fremy
}
