% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_pca.R, R/plot_pca.R
\name{mod_ds_pca_server}
\alias{mod_ds_pca_server}
\alias{wrapper_pca}
\alias{plotPCA_Eigen}
\title{#' @param res.pca Result of FactoMineR::PCA
#' @param chosen.axes The dimensions to plot
#'
#' @author Samuel Wieczorek, Enora Fremy
#'
#' @importFrom factoextra fviz_pca_var
#'
#' @export
#'
#' @rdname ds-pca
#'
plotPCA_Var <- function(res.pca = NULL,
chosen.axes = c(1, 2)){
if (is.null(res.pca))
return(NULL)

fviz_pca_var(res.pca,
axes = chosen.axes,
col.var = "cos2",
gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"),
repel = TRUE
)
}
#' @param res.pca Result of FactoMineR::PCA
#' @param axes The dimensions to plot
#' @param geom xxx
#'
#' @author Samuel Wieczorek, Enora Fremy
#'
#' @importFrom factoextra fviz_pca_ind
#'
#' @export
#'
#' @rdname ds-pca
#'
plotPCA_Ind <- function(res.pca, chosen.axes = c(1,2)){
if (is.null(res.pca))
return(NULL)

fviz_pca_ind(res.pca,
axes = chosen.axes,
geom  ="point")

}}
\usage{
mod_ds_pca_server(id, data, conds)

wrapper_pca(data, conds, var.scaling = TRUE, ncp = NULL)

plotPCA_Eigen(res.pca)
}
\arguments{
\item{id}{xxx.}

\item{data}{An instance of the class \code{SummarizedExperiment}}

\item{conds}{xxx}

\item{var.scaling}{The dimensions to plot}

\item{ncp}{A \code{integer(1)} which represents the umber of dimensions kept in the results.}

\item{res.pca}{Result of FactoMineR::PCA}
}
\description{
#' @param res.pca Result of FactoMineR::PCA
#' @param chosen.axes The dimensions to plot
#'
#' @author Samuel Wieczorek, Enora Fremy
#'
#' @importFrom factoextra fviz_pca_var
#'
#' @export
#'
#' @rdname ds-pca
#'
plotPCA_Var <- function(res.pca = NULL,
chosen.axes = c(1, 2)){
if (is.null(res.pca))
return(NULL)

fviz_pca_var(res.pca,
axes = chosen.axes,
col.var = "cos2",
gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"),
repel = TRUE
)
}
#' @param res.pca Result of FactoMineR::PCA
#' @param axes The dimensions to plot
#' @param geom xxx
#'
#' @author Samuel Wieczorek, Enora Fremy
#'
#' @importFrom factoextra fviz_pca_ind
#'
#' @export
#'
#' @rdname ds-pca
#'
plotPCA_Ind <- function(res.pca, chosen.axes = c(1,2)){
if (is.null(res.pca))
return(NULL)

fviz_pca_ind(res.pca,
axes = chosen.axes,
geom  ="point")

}
}
\author{
Samuel Wieczorek, Enora Fremy
}
