% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_pca.R, R/plot_pca.R
\name{ds-pca}
\alias{ds-pca}
\alias{mod_ds_pca_ui}
\alias{mod_ds_pca_server}
\alias{wrapper_pca}
\alias{plotPCA_Eigen}
\title{Bar plot of missing values per lines using highcharter.}
\usage{
mod_ds_pca_ui(id)

mod_ds_pca_server(id, vizData)

wrapper_pca(vizData, var.scaling = TRUE, ncp = NULL)

plotPCA_Eigen(res.pca)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{var.scaling}{The dimensions to plot}

\item{ncp}{A \code{integer(1)} which represents the umber of dimensions kept in
the results.}

\item{res.pca}{Result of FactoMineR::PCA}

\item{data}{An instance of the class \code{SummarizedExperiment} which is the
quantitative values.}

\item{conds}{A \code{character()} of the name of conditions
(one condition per sample). The number of conditions must be equal to
the number of samples (number of columns) of the parameter 'data'.}
}
\value{
A plot
}
\description{
This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
}
\examples{

}
\author{
Samuel Wieczorek, Enora Fremy
}
