% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_pca.R, R/plot_pca.R
\name{ds-pca}
\alias{ds-pca}
\alias{mod_ds_pca_ui}
\alias{mod_ds_pca_server}
\alias{wrapper_pca}
\alias{plotPCA_Eigen}
\title{Bar plot of missing values per lines using highcharter.}
\usage{
mod_ds_pca_ui(id)

mod_ds_pca_server(id, vizData)

wrapper_pca(vizData, var.scaling = TRUE, ncp = NULL)

plotPCA_Eigen(res.pca)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{vizData}{An instance of the class \code{VizData}.}

\item{var.scaling}{The dimensions to plot}

\item{ncp}{A \code{integer(1)} which represents the umber of dimensions kept in
the results.}

\item{res.pca}{Result of FactoMineR::PCA}
}
\value{
A plot
}
\description{
This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
}
\examples{
library(highcharter)
library(DaparViz)
library(shiny)

data(Exp1_R25_pept, package='DaparToolshedData')
vList <- convert2viz(Exp1_R25_pept)
vData <- vList@ll.vizData[[1]]


res.pca <- wrapper_pca(vData, ncp = 6)
plotPCA_Var(res.pca)
plotPCA_Ind(res.pca)
plotPCA_Eigen_hc(res.pca)




ui <- mod_ds_pca_ui("plot")

server <- function(input, output, session) {
  data(Exp1_R25_pept, package='DaparToolshedData')
  vList <- convert2viz(Exp1_R25_pept)
  vData <- vList@ll.vizData[[1]]
  
  mod_ds_pca_server("plot", vizData = reactive({vData}))
  }

if (interactive())
  shinyApp(ui = ui, server = server)
library(highcharter)
library(DaparViz)
library(shiny)

data(Exp1_R25_pept, package='DaparToolshedData')
vList <- convert2viz(Exp1_R25_pept)
vData <- vList@ll.vizData[[1]]


res.pca <- wrapper_pca(vData, ncp = 6)
plotPCA_Var(res.pca)
plotPCA_Ind(res.pca)
plotPCA_Eigen_hc(res.pca)




ui <- mod_ds_pca_ui("plot")

server <- function(input, output, session) {
  data(Exp1_R25_pept, package='DaparToolshedData')
  vList <- convert2viz(Exp1_R25_pept)
  vData <- vList@ll.vizData[[1]]
  
  mod_ds_pca_server("plot", vizData = reactive({vData}))
  }

if (interactive())
  shinyApp(ui = ui, server = server)
library(highcharter)
library(DaparViz)
library(shiny)

data(Exp1_R25_pept, package='DaparToolshedData')
vList <- convert2viz(Exp1_R25_pept)
vData <- vList@ll.vizData[[1]]


res.pca <- wrapper_pca(vData, ncp = 6)
plotPCA_Var(res.pca)
plotPCA_Ind(res.pca)
plotPCA_Eigen_hc(res.pca)




ui <- mod_ds_pca_ui("plot")

server <- function(input, output, session) {
  data(Exp1_R25_pept, package='DaparToolshedData')
  vList <- convert2viz(Exp1_R25_pept)
  vData <- vList@ll.vizData[[1]]
  
  mod_ds_pca_server("plot", vizData = reactive({vData}))
  }

if (interactive())
  shinyApp(ui = ui, server = server)
}
\author{
Samuel Wieczorek, Enora Fremy
}
