% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_pca.R, R/plot_pca.R
\name{ds-pca}
\alias{ds-pca}
\alias{mod_ds_pca_ui}
\alias{mod_ds_pca_server}
\alias{ds_pca}
\alias{wrapper_pca}
\alias{plotPCA_Eigen}
\alias{plotPCA_Var}
\alias{plotPCA_Ind}
\alias{plotPCA_Eigen_hc}
\title{Bar plot of missing values per lines using highcharter.}
\usage{
mod_ds_pca_ui(id)

mod_ds_pca_server(id, obj)

ds_pca(obj)

wrapper_pca(DaparViz, var.scaling = TRUE, ncp = NULL)

plotPCA_Eigen(res.pca)

plotPCA_Var(res.pca, chosen.axes = c(1, 2))

plotPCA_Ind(res.pca, chosen.axes = c(1, 2))

plotPCA_Eigen_hc(res.pca)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An instance of the class \code{DaparViz}}

\item{DaparViz}{An instance of the class \code{DaparViz}.}

\item{var.scaling}{The dimensions to plot}

\item{ncp}{A \code{integer(1)} which represents the umber of dimensions kept in
the results.}

\item{res.pca}{Result of FactoMineR::PCA}

\item{chosen.axes}{The dimensions to plot}
}
\value{
A plot

A plot
}
\description{
This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).

xxx
\itemize{
\item \code{wrapper_pca()}: xxx
\item \code{plotPCA_Eigen_hc()}: plots the eigen values of PCA with the highcharts library
\item \code{plotPCA_Eigen()}: plots the eigen values of PCA
\item \code{plotPCA_Var()}:
\item \code{plotPCA_Ind()}:
}
}
\examples{
if(interactive()){
data(vData_ft)
ds_pca(vData_ft[[1]])
}

}
\author{
Samuel Wieczorek, Enora Fremy
}
