% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_pca.R, R/plot_pca.R
\name{ds-pca}
\alias{ds-pca}
\alias{omXplore_pca_ui}
\alias{omXplore_pca_server}
\alias{omXplore_pca}
\alias{wrapper_pca}
\alias{plotPCA_Eigen}
\alias{plotPCA_Var}
\alias{plotPCA_Ind}
\alias{plotPCA_Eigen_hc}
\title{Bar plot of missing values per lines using highcharter.}
\usage{
omXplore_pca_ui(id)

omXplore_pca_server(id, obj)

omXplore_pca(obj)

wrapper_pca(obj, var.scaling = TRUE, ncp = NULL)

plotPCA_Eigen(res.pca)

plotPCA_Var(res.pca, chosen.axes = c(1, 2))

plotPCA_Ind(res.pca, chosen.axes = c(1, 2))

plotPCA_Eigen_hc(res.pca)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{xxx}

\item{var.scaling}{xxx}

\item{ncp}{xxx}

\item{res.pca}{xxx}

\item{chosen.axes}{xxx}
}
\value{
NA

NA

A shiny app

The result of the function FactoMineR::PCA()

A plot

A plot

A plot

A plot
}
\description{
This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
\itemize{
\item \code{wrapper_pca()}: xxx
\item \code{plotPCA_Eigen_hc()}: plots the eigen values of PCA with the highcharts
library
\item \code{plotPCA_Eigen()}: plots the eigen values of PCA
\item \code{plotPCA_Var()}:
\item \code{plotPCA_Ind()}:
}

xxx
}
\examples{
if (interactive()) {
  data(vData_ft)
  obj <- vData_ft[[1]]
  # Replace missing values for the example
  qdata <- GetSlotQdata(obj)
  qdata[which(is.na(qdata))] <- 0
  omXplore_pca(obj)
}

data(vData_ft)
obj <- vData_ft[[1]]
res.pca <- wrapper_pca(obj)
plotPCA_Eigen(res.pca)
plotPCA_Var(res.pca)
plotPCA_Eigen_hc(res.pca)
plotPCA_Ind(res.pca)

}
\author{
Samuel Wieczorek, Enora Fremy
}
