% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_view_dataset.R
\name{ds-plots}
\alias{ds-plots}
\alias{listPlotModules}
\alias{mod_view_dataset_ui}
\alias{mod_view_dataset_server}
\title{Bar plot of missing values per lines using highcharter.}
\usage{
listPlotModules()

mod_view_dataset_ui(id)

mod_view_dataset_server(
  id,
  ll.vizData = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{A \code{character(1)} for the 'id' of the shiny module. It must be
the same as for the '*_ui' function.}

\item{ll.vizData}{A instance of the class \code{VizList}.}
}
\value{
A plot
}
\description{
This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
}
\details{
\itemize{
\item distribution of the missing values per line,
\item a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins) and per conditions,
\item Histogram of missing values.
\item Variance : Builds a densityplot of the CV of entities in numeric matrix.
The CV is calculated for each condition present in the dataset
(see the slot \code{'Condition'} in the \code{colData()} DataFrame)
\item Heatmap:
}

The function \code{\link[=heatmapD]{heatmapD()}}

The function [] is inspired from the function 'heatmap.2'
that displays a numeric matrix. For more information, please refer to the help
of the heatmap.2 function.
}
\section{Missing values}{


#' - distribution of the missing values per line,
\itemize{
\item a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins) and per conditions,
\item Histogram of missing values.
}
}

\examples{
library(DaparViz)
library(highcharter)
library(MSnbase)
library(shiny)


options(shiny.fullstacktrace = TRUE)



ui <- fluidPage(
  mod_view_dataset_ui("dataset")
)



server <- function(input, output, session) {
   
  #
  # Example with a QFeatures dataset
  #
  data(ft)
  obj <- ft
  
  #
  # Example with a series of MSnSet datasets
  #
  # data(Exp1_R25_prot, package='DAPARdata')
  # data(Exp1_R25_pept, package='DAPARdata')
  # data(Exp1_R2_pept, package='DAPARdata')
  # obj <- setNames(c(Exp1_R25_prot, Exp1_R25_pept, Exp1_R2_pept),
  #                    nm = c('Exp1_R25_prot', 'Exp1_R25_pept', 'Exp1_R2_pept'))
  # 
  # 
  
  mod_view_dataset_server("dataset", 
                          ll.vizData = reactive({convert2viz(obj)}))
}

if (interactive())
  shinyApp(ui, server)
library(DaparViz)
library(highcharter)
library(MSnbase)
library(shiny)


options(shiny.fullstacktrace = TRUE)



ui <- fluidPage(
  mod_view_dataset_ui("dataset")
)



server <- function(input, output, session) {
   
  #
  # Example with a QFeatures dataset
  #
  data(ft)
  obj <- ft
  
  #
  # Example with a series of MSnSet datasets
  #
  # data(Exp1_R25_prot, package='DAPARdata')
  # data(Exp1_R25_pept, package='DAPARdata')
  # data(Exp1_R2_pept, package='DAPARdata')
  # obj <- setNames(c(Exp1_R25_prot, Exp1_R25_pept, Exp1_R2_pept),
  #                    nm = c('Exp1_R25_prot', 'Exp1_R25_pept', 'Exp1_R2_pept'))
  # 
  # 
  
  mod_view_dataset_server("dataset", 
                          ll.vizData = reactive({convert2viz(obj)}))
}

if (interactive())
  shinyApp(ui, server)
}
\author{
Samuel Wieczorek, Enora Fremy
}
