% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds.R
\name{ds-plots}
\alias{ds-plots}
\alias{listPlotModules}
\alias{mod_ds_ui}
\alias{mod_ds_server}
\title{Bar plot of missing values per lines using highcharter.}
\usage{
listPlotModules()

mod_ds_ui(id)

mod_ds_server(id, object)
}
\arguments{
\item{id}{A \code{character(1)} for the 'id' of the shiny module. It must be
the same as for the '*_ui' function.}

\item{object}{A instance of the class \code{QFeatures}.}
}
\value{
A plot
}
\description{
This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
}
\details{
\itemize{
\item distribution of the missing values per line,
\item a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins) and per conditions,
\item Histogram of missing values.
\item Variance : Builds a densityplot of the CV of entities in numeric matrix.
The CV is calculated for each condition present in the dataset
(see the slot \code{'Condition'} in the \code{colData()} DataFrame)
\item Heatmap:
}

The function \code{\link[=heatmapD]{heatmapD()}}

The function [] is inspired from the function 'heatmap.2'
that displays a numeric matrix. For more information, please refer to the help
of the heatmap.2 function.
}
\section{Missing values}{


#' - distribution of the missing values per line,
\itemize{
\item a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins) and per conditions,
\item Histogram of missing values.
}
}

\examples{
library(QFeatures)
library(DaparToolshed)
data(ft)
data(ft_na)

qData <- assay(ft, 1)
conds <- design(ft)$Condition
pal <- 'Dark2'

data(ft)
qData <- assay(ft, 1)
conds <- design(ft)$Condition
densityPlot(qData, conds)

legend <- design(ft)$Sample.name
densityPlot(qData, conds, pal.name = 'Dark2')

CVDist(qData, conds)

#----------------------------------------------
# Plots a heatmap for generic quantitative data
#----------------------------------------------

heatmapD(assay(ft, 1), conds)

#----------------------------------------------
# Plots a heatmap for missing values visualization
#----------------------------------------------


#' #----------------------------------------
# Plots a histogram of missing values
#----------------------------------------

mvPerLinesHisto(qData)

#----------------------------------------
# Plots a histogram of missing values
#----------------------------------------

mvHisto(qData, conds, pal.name = pal)

#----------------------------------------
# Plots a histogram of missing values
#----------------------------------------



library(QFeatures)
data(ft)
mv.heatmap(assay(ft, 1))

#----------------------------------------
# Launch a the shiny module for all plots
# Here, an example with the density plot
#----------------------------------------

if(interactive()){
 data(ft_na)
 ui <- mod_ds_ui('plot')

 server <- function(input, output, session) {
  mod_ds_server('plot', object = reactive({ft_na}))
  }
 
 shinyApp(ui=ui, server=server)
}
}
\author{
Samuel Wieczorek, Enora Fremy
}
