% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_apps_examples.R
\name{external_app}
\alias{external_app}
\alias{extFoo1_ui}
\alias{extFoo1_server}
\alias{extFoo1}
\alias{extFoo2_ui}
\alias{extFoo2_server}
\alias{extFoo2}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
extFoo1_ui(id)

extFoo1_server(
  id,
  obj = reactive({
     NULL
 }),
  i = reactive({
     1
 })
)

extFoo1(obj)

extFoo2_ui(id)

extFoo2_server(
  id,
  obj = reactive({
     NULL
 }),
  i = reactive({
     1
 })
)

extFoo2(obj)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An object of instance \code{MultiAssayExperiment}}

\item{i}{xxx}
}
\value{
NA

NA

NA

A shiny app

NA

NA

A shiny app
}
\description{
xxxx
}
\examples{
if (interactive()) {
  data(vdata)
  extFoo1(vdata, 1)
  extFoo2(vdata, 1)
}

}
