% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extFoo1.R, R/extFoo2.R
\name{external_app}
\alias{external_app}
\alias{extFoo1_ui}
\alias{extFoo2_ui}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
extFoo1_ui(id)

extFoo2_ui(id)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An instance of the class \code{DaparViz}}
}
\value{
NA

NA
}
\description{
xxxx

xxxx
}
\examples{
if(interactive()){
data(vData_ft)
DaparViz_corrmatrix(vData_ft[[1]])
}


if(interactive()){
data(vData_ft)
DaparViz_corrmatrix(vData_ft[[1]])
}


}
