% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_funcs.R
\name{FormatAvailables}
\alias{FormatAvailables}
\title{Formats available in DaparViz}
\usage{
FormatAvailables()
}
\value{
A vector of classes compliant with the package DaparViz.
}
\description{
Gives the list of classes for the datasets which be can converted by the
function \code{convert2Viz()}.
}
\examples{
FormatAvailables()

}
