% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FormatDataForDT}
\alias{FormatDataForDT}
\title{Constructs a dataset suitable to use with the module format_DT.}
\usage{
FormatDataForDT(vizData, digits = 2)
}
\arguments{
\item{vizData}{An instance of the class \code{DaparViz}}

\item{digits}{An 'integer(1)' to specify the number of digits to display
in the tables for numerical values. Default is 2.}
}
\value{
A data.frame
}
\description{
This function builds the skeleton of a dataset which can be used by the module
format_DT. It creates additional columns to be used to style the table.
to colors cells.
}
