% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{heatmapForMissingValues}
\alias{heatmapForMissingValues}
\title{xxx}
\usage{
heatmapForMissingValues(
  x,
  col = NULL,
  srtCol = NULL,
  labCol = NULL,
  labRow = NULL,
  key = TRUE,
  key.title = NULL,
  main = NULL,
  ylab = NULL
)
}
\arguments{
\item{x}{A dataframe that contains quantitative data.}

\item{col}{colors used for the image. Defaults to heat colors (heat.colors).}

\item{srtCol}{angle of column conds, in degrees from horizontal}

\item{labCol}{character vectors with column conds to use.}

\item{labRow}{character vectors with row conds to use.}

\item{key}{logical indicating whether a color-key should be shown.}

\item{key.title}{main title of the color key. If set to NA no title will
be plotted.}

\item{main}{main title; default to none.}

\item{ylab}{y-axis title; default to none.}
}
\value{
A heatmap
}
\description{
This function is inspired from the function \code{heatmap.2}
that displays quantitative data in the \code{MSnbase::exprs()} table of an
object of
class \code{MSnSet}. For more information, please refer to the help
of the heatmap.2 function.
}
\examples{
library(shiny)


ui <- fluidPage(
  mod_ds_heatmap_ui("plot")
)

server <- function(input, output, session) {
  
  data(vData_ft)
  obj <- vData_ft[1]
  
  mod_ds_heatmap_server("plot", reactive({obj})
  )
  }

if (interactive())
  shinyApp(ui = ui, server = server)
}
\author{
Samuel Wieczorek
}
