% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_heatmap.R, R/plot_heatmap.R
\name{heatmaps}
\alias{heatmaps}
\alias{mod_ds_heatmap_ui}
\alias{mod_ds_heatmap_server}
\alias{mv.heatmap}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_heatmap_ui(id)

mod_ds_heatmap_server(
  id,
  vizData = reactive({
     NULL
 }),
  width = 900
)

mv.heatmap(
  x,
  col = grDevices::heat.colors(100),
  srtCol = NULL,
  labCol = NULL,
  labRow = NULL,
  key = TRUE,
  key.title = NULL,
  main = NULL,
  ylab = NULL
)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{vizData}{xxx}

\item{width}{xxx}

\item{x}{A \code{matrix} or \code{array} containing the quantitative data.}

\item{col}{Colors used for the image. Defaults to heat colors (heat.colors).}

\item{srtCol}{Angle of column conds, in degrees from horizontal}

\item{labCol}{Character vectors with column conds to use.}

\item{labRow}{Character vectors with row conds to use.}

\item{key}{Logical indicating whether a color-key should be shown.}

\item{key.title}{Main title of the color key. If set to NA no title will
be plotted.}

\item{main}{Main title; default to none.}

\item{ylab}{y-axis title; default to none.}
}
\value{
NA
}
\description{
xxxx
}
