% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_intensity.R, R/plot_violin.R
\name{intensity-plots}
\alias{intensity-plots}
\alias{omXplore_intensity_ui}
\alias{omXplore_intensity_server}
\alias{omXplore_intensity}
\alias{violinPlot}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
omXplore_intensity_ui(id)

omXplore_intensity_server(
  id,
  obj,
  i,
  track.indices = reactive({
     NULL
 })
)

omXplore_intensity(obj, i, withTracking = FALSE)

violinPlot(data, conds, subset = NULL, pal.name)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{A instance of the class \code{MultiAssayExperiment}}

\item{i}{xxx}

\item{track.indices}{xxx}

\item{withTracking}{xxx}

\item{data}{xxx}

\item{conds}{xxx}

\item{subset}{xxx}

\item{pal.name}{xxx}
}
\value{
NA

NA

A shiny app

A violin plot
}
\description{
xxxx
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
  data(vdata)
  shiny::runApp(omXplore_intensity(vdata, 1))
\dontshow{\}) # examplesIf}
}
