% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_intensity.R, R/plot_boxplot.R,
%   R/plot_violin.R
\name{intensity-plots}
\alias{intensity-plots}
\alias{omXplore_intensity_ui}
\alias{omXplore_intensity_server}
\alias{omXplore_intensity}
\alias{boxPlot}
\alias{violinPlot}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
omXplore_intensity_ui(id)

omXplore_intensity_server(
  id,
  obj,
  i,
  track.indices = reactive({
     NULL
 })
)

omXplore_intensity(obj, i, withTracking = FALSE)

boxPlot(data, conds, subset = NULL, pal.name = NULL)

violinPlot(data, conds, subset = NULL, pal.name)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{A instance of the class \code{MultiAssayExperiment}}

\item{i}{xxx}

\item{track.indices}{xxx}

\item{withTracking}{xxx}

\item{subset}{A \code{integer()} vector of index indicating the indices
of rows in the dataset to highlight}

\item{pal.name}{A \code{character(1)} which is the name of the palette from the
package \link{RColorBrewer} from which the colors are taken. Default value
is 'Set1'.}
}
\value{
NA

NA

A shiny app

A boxplot

A violin plot
}
\description{
xxxx
}
\examples{
if (interactive()) {
  data(vdata)
  omXplore_intensity(vdata, 1)
}

data(vdata)
qdata <- GetSlotQdata(vdata, 1)
conds <- GetSlotConds(vdata, 1)
boxPlot(qdata, conds)

}
\author{
Samuel Wieczorek, Anais Courtier, Enora Fremy
}
