% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_intensity.R, R/plot_boxplot.R,
%   R/plot_violin.R
\name{intensity-plots}
\alias{intensity-plots}
\alias{mod_ds_intensity_ui}
\alias{mod_ds_intensity_server}
\alias{boxPlot}
\alias{violinPlot}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_intensity_ui(id)

mod_ds_intensity_server(id, vizData)

boxPlot(data, conds, subset = NULL, pal.name = NULL)

violinPlot(data, conds, subset = NULL, pal.name)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{vizData}{A instance of the class \code{VizData}}

\item{data}{An instance of the class \code{matrix}}

\item{conds}{A \code{character()} of the name of conditions
(one condition per sample). The number of conditions must be equal to
the number of samples (number of columns) of the parameter 'data'.}

\item{subset}{A vector of index indicating rows to highlight}

\item{pal.name}{A \code{character(1)} which is the name of the palette from the
package \link{RColorBrewer} from which the colors are taken. Default value
is 'Set1'.}
}
\value{
NA

NA

A boxplot

A violinplot
}
\description{
xxxx
}
\author{
Samuel Wieczorek, Anais Courtier, Enora Fremy
}
