% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparViz_intensity.R, R/plot_boxplot.R,
%   R/plot_violin.R
\name{intensity-plots}
\alias{intensity-plots}
\alias{DaparViz_intensity_ui}
\alias{DaparViz_intensity_server}
\alias{DaparViz_intensity}
\alias{boxPlot}
\alias{violinPlot}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
DaparViz_intensity_ui(id)

DaparViz_intensity_server(
  id,
  obj,
  track.indices = reactive({
     NULL
 })
)

DaparViz_intensity(obj, withTracking = FALSE)

boxPlot(data, conds, subset = NULL, pal.name = NULL)

violinPlot(data, conds, subset = NULL, pal.name)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{A instance of the class \code{DaparViz}}

\item{track.indices}{xxx}

\item{withTracking}{xxx}

\item{data}{Numeric matrix}

\item{conds}{A \code{character()} of the name of conditions
(one condition per sample). The number of conditions must be equal to
the number of samples (number of columns) of the parameter 'data'.}

\item{subset}{A \code{integer()} vector of index indicating the indices
of rows in the dataset to highlight}

\item{pal.name}{A \code{character(1)} which is the name of the palette from the
package \link{RColorBrewer} from which the colors are taken. Default value
is 'Set1'.}
}
\value{
NA

NA

A shiny app

A boxplot

A violin plot
}
\description{
xxxx
}
\examples{
if (interactive()) {
  data(vData_ft)
  DaparViz_intensity(vData_ft[[1]])
}

data(vData_ft)
qdata <- GetSlotQdata(vData_ft[[1]])
conds <- GetSlotConds(vData_ft[[1]])
boxPlot(qdata, conds)

}
\author{
Samuel Wieczorek, Anais Courtier, Enora Fremy
}
