% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_intensity.R, R/plot_boxplot.R,
%   R/plot_violin.R
\name{intensity-plots}
\alias{intensity-plots}
\alias{mod_ds_intensity_ui}
\alias{mod_ds_intensity_server}
\alias{boxPlot}
\alias{violinPlot}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_intensity_ui(id)

mod_ds_intensity_server(id, se, conds, ...)

boxPlot(data, conds, subset = NULL, pal.name = NULL)

violinPlot(data, conds, subset = NULL, pal.name)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{se}{A instance of the class \code{SummarizedExperiment}}

\item{conds}{A \code{character()} of the name of conditions
(one condition per sample). The number of conditions must be equal to
the number of samples (number of columns) of the parameter 'data'.}

\item{...}{Additional parameters for \code{\link[=boxPlot]{boxPlot()}} or \code{\link[=violinPlot]{violinPlot()}}.}

\item{data}{An instance of the class \code{matrix}}

\item{subset}{A vector of index indicating rows to highlight}

\item{pal.name}{A \code{character(1)} which is the name of the palette from the
package \link{RColorBrewer} from which the colors are taken. Default value
is 'Set1'.}
}
\value{
NA

NA

A boxplot

A violinplot
}
\description{
xxxx
}
\examples{

library(SummarizedExperiment)
data(ft, package='DaparViz')
conds <- colData(ft)$Condition
violinPlot(assay(ft, 1), conds)
violinPlot(assay(ft, 1), conds, subset = c(2, 4))

boxPlot(assay(ft, 1), conds)
boxPlot(assay(ft, 1), conds, subset = c(2, 4))


#------------------------------------------
# Shiny module
#------------------------------------------

    data(ft, package='DaparViz')
    ui <- mod_ds_intensity_ui("iplot")

    server <- function(input, output, session) {
        mod_ds_intensity_server("iplot", reactive({ft[[1]]}),
            colData(ft)$Condition)
    }

    shinyApp(ui = ui, server = server)
}
\author{
Samuel Wieczorek, Anais Courtier, Enora Fremy
}
