% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.listOf}
\alias{is.listOf}
\title{xxx}
\usage{
is.listOf(object, obj.class = NULL)
}
\arguments{
\item{object}{A \code{list}}

\item{obj.class}{The name of the class to search in items of the list.}
}
\value{
A \code{character(1)} with the name of the package or xxx
}
\description{
xxx
}
\examples{
NULL

}
