% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_metacell.R, R/metacellPlots.R
\name{metacell-plots}
\alias{metacell-plots}
\alias{mod_ds_metacell_ui}
\alias{mod_ds_mv_server}
\alias{metacellPerLinesHisto_HC}
\alias{metacellPerLinesHistoPerCondition_HC}
\alias{metacellHisto_HC}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_metacell_ui(id)

mod_ds_mv_server(
  id,
  obj = reactive({
     NULL
 }),
  design = NULL,
  pal = reactive({
     NULL
 }),
  pattern = reactive({
     NULL
 }),
  showSelect = reactive({
     TRUE
 })
)

metacellPerLinesHisto_HC(
  object,
  design,
  pattern = NULL,
  detailed = FALSE,
  indLegend = "auto",
  showValues = FALSE
)

metacellPerLinesHistoPerCondition_HC(
  object,
  design,
  pattern = NULL,
  indLegend = "auto",
  showValues = FALSE,
  pal = NULL
)

metacellHisto_HC(
  object,
  design,
  pattern = NULL,
  indLegend = "auto",
  showValues = FALSE,
  pal = NULL
)
}
\arguments{
\item{design}{xxx}

\item{pattern}{xxx}

\item{object}{An instance of the class \code{SummarizedExperiment}}

\item{detailed}{'value' or 'percent'}

\item{indLegend}{The indices of the column name's in \code{Biobase::pData()}
tab}

\item{showValues}{A logical that indicates whether numeric values should be
drawn above the bars.}
}
\value{
NA

A bar plot
}
\description{
xxxx

This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
}
\examples{

library(SummarizedExperiment)
data(ft_na, package='DaparViz')
mvPerLinesHisto(assay(ft_na, 1))

mvPerLinesHistoPerCondition(assay(ft_na, 1), colData(ft_na)$Condition)
mvHisto(assay(ft_na, 1), colData(ft_na)$Condition)

#------------------------------------------
# Shiny module
#------------------------------------------

data(ft_na, package='DaparViz')
ui <- mod_ds_mv_ui("plot")

server <- function(input, output, session) {
  mod_ds_mv_server("plot",
                   se = reactive({assay(ft_na, 1)}),
                   design = design(ft_na))
    }

shinyApp(ui = ui, server = server)
data(Exp1_R25_pept, package="DAPARdata")
obj <- Exp1_R25_pept
pattern <- "Missing POV"
pal <- ExtendPalette(2, "Dark2")
metacellHisto_HC(obj, pattern, showValues = TRUE, pal = pal)

}
\author{
Florence Combes, Samuel Wieczorek
}
