% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_metacell.R, R/metacellPlots.R
\name{metacell-plots}
\alias{metacell-plots}
\alias{mod_ds_metacell_ui}
\alias{mod_ds_mv_server}
\alias{metacellPerLinesHisto_HC}
\alias{metacellPerLinesHistoPerCondition_HC}
\alias{metacellHisto_HC}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_metacell_ui(id)

mod_ds_mv_server(
  id,
  obj = reactive({
     NULL
 }),
  design = NULL,
  pal = reactive({
     NULL
 }),
  pattern = reactive({
     NULL
 }),
  showSelect = reactive({
     TRUE
 })
)

metacellPerLinesHisto_HC(
  qdata,
  metacell = NULL,
  type,
  design,
  pattern = NULL,
  detailed = FALSE,
  indLegend = "auto",
  showValues = FALSE
)

metacellPerLinesHistoPerCondition_HC(
  qdata,
  metacell = NULL,
  type,
  design,
  pattern = NULL,
  indLegend = "auto",
  showValues = FALSE,
  pal = NULL
)

metacellHisto_HC(
  qdata,
  metacell = NULL,
  type,
  design,
  pattern = NULL,
  indLegend = "auto",
  showValues = FALSE,
  pal = NULL
)
}
\arguments{
\item{design}{xxx}

\item{pattern}{xxx}

\item{qdata}{An instance of the class \code{SummarizedExperiment}}

\item{metacell}{xxx}

\item{type}{xxx}

\item{detailed}{'value' or 'percent'}

\item{indLegend}{The indices of the column name's in \code{Biobase::pData()}
tab}

\item{showValues}{A logical that indicates whether numeric values should be
drawn above the bars.}
}
\value{
NA

A bar plot
}
\description{
xxxx

This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
}
\examples{

library(SummarizedExperiment)
data(ft_na, package='DaparViz')
mvPerLinesHisto(assay(ft_na, 1))

mvPerLinesHistoPerCondition(assay(ft_na, 1), colData(ft_na)$Condition)
mvHisto(assay(ft_na, 1), colData(ft_na)$Condition)

#------------------------------------------
# Shiny module
#------------------------------------------

data(ft_na, package='DaparViz')

ui <- fluidPage(
  mod_ds_metacell_ui('test')
)

server <- function(input, output) {
  utils::data("Exp1_R25_prot", package='DaparToolshedData')
  obj <- Exp1_R25_prot
  pattern <- c('Missing POV', 'Missing MEC')
  
  # pattern <- NULL
  observe({
    mod_ds_metacell_server('test',
                           obj = reactive({obj[[1]]}),
                           design = design.qf(obj),
                           pal = reactive({NULL}),
                           pattern = reactive({pattern}),
                           showSelect = reactive({is.null(pattern)})
    )
  })
  
}

shinyApp(ui = ui, server = server)

data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept

obj <- obj[1:10]
se <- obj[[1]]

qdata <- assay(obj[[1]])
metacell <- qMetacell(obj[[1]])
type <- typeDataset(obj[[1]])
design <- colData(obj)


pal <- ExtendPalette(length(unique(design$Condition)), "Dark2")


metacellPerLinesHisto_HC(qdata, metacell, type, design = design, pattern = "Missing POV")
metacellPerLinesHisto_HC(qdata, metacell, type, design = design)
metacellPerLinesHisto_HC(qdata, metacell, type, design = design, pattern = "Quantified")
metacellPerLinesHisto_HC(qdata, metacell, type, design = design, pattern = "Quant. by direct id")
metacellPerLinesHisto_HC(qdata, metacell, type, design = design, pattern = "Quant. by recovery")
metacellPerLinesHisto_HC(qdata, metacell, type, design = design, pattern = c("Quantified", "Quant. by direct id", "Quant. by recovery"))


metacellPerLinesHistoPerCondition_HC(qdata, metacell, type, design = design, pattern = "Missing POV")
metacellPerLinesHistoPerCondition_HC(qdata, metacell, type, design = design)
metacellPerLinesHistoPerCondition_HC(qdata, metacell, type, design = design, pattern = "Quantified")
metacellPerLinesHistoPerCondition_HC(qdata, metacell, type, design = design, pattern = "Quant. by direct id")
metacellPerLinesHistoPerCondition_HC(qdata, metacell, type, design = design, pattern = "Quant. by recovery")
metacellPerLinesHistoPerCondition_HC(qdata, metacell, type, design = design, pattern = c("Quantified", "Quant. by direct id", "Quant. by recovery"))


metacellHisto_HC(qdata, metacell, type, design = design, pattern = "Missing POV")
metacellHisto_HC(qdata, metacell, type, design = design)
metacellHisto_HC(qdata, metacell, type, design = design, pattern = "Quantified")
metacellHisto_HC(qdata, metacell, type, design = design, pattern = "Quant. by direct id")
metacellHisto_HC(qdata, metacell, type, design = design, pattern = "Quant. by recovery")
metacellHisto_HC(qdata, metacell, type, design = design, pattern = c("Quantified", "Quant. by direct id", "Quant. by recovery"))
data(Exp1_R25_pept, package="DAPARdata")
obj <- Exp1_R25_pept
pattern <- "Missing POV"
pal <- ExtendPalette(2, "Dark2")
metacellHisto_HC(obj, pattern, showValues = TRUE, pal = pal)

}
\author{
Florence Combes, Samuel Wieczorek
}
