% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_metacell_tree.R
\name{metacell-tree}
\alias{metacell-tree}
\alias{mod_metacell_tree_ui}
\alias{mod_metacell_tree_server}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_metacell_tree_ui(id)

mod_metacell_tree_server(
  id,
  obj = reactive({
     NULL
 }),
  reset = reactive({
     NULL
 })
)
}
\value{
NA
}
\description{
xxxx
}
\examples{

library(SummarizedExperiment)
data(ft_na, package='DaparViz')
mvPerLinesHisto(assay(ft_na, 1))

mvPerLinesHistoPerCondition(assay(ft_na, 1), colData(ft_na)$Condition)
mvHisto(assay(ft_na, 1), colData(ft_na)$Condition)

#------------------------------------------
# Shiny module
#------------------------------------------

data(ft_na, package='DaparViz')
ui <- mod_ds_mv_ui("plot")

server <- function(input, output, session) {
  mod_ds_mv_server("plot",
                   se = reactive({assay(ft_na, 1)}),
                   design = design(ft_na))
    }

shinyApp(ui = ui, server = server)
}
