% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_metacell_tree.R
\name{metacell-tree}
\alias{metacell-tree}
\alias{mod_metacell_tree_ui}
\alias{mod_metacell_tree_server}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_metacell_tree_ui(id)

mod_metacell_tree_server(
  id,
  type = reactive({
     NULL
 }),
  reset = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{xxx}

\item{type}{xxx}

\item{reset}{xxx}
}
\value{
NA
}
\description{
xxxx
}
\examples{

# Example
#
ui <- fluidPage(
  mod_metacell_tree_ui('tree')
)

server <- function(input, output) {
  rv <- reactiveValues(
    tags = NULL
  )
  
  observe({
    rv$tags <- mod_metacell_tree_server('tree', type = reactive({'peptide'}))
  })
  
  
}

shinyApp(ui = ui, server = server)
}
