% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_format_DT.R
\name{mod_format_DT}
\alias{mod_format_DT}
\alias{mod_format_DT_ui}
\alias{mod_format_DT_server}
\title{mod_format_DT_ui and mod_format_DT_server}
\usage{
mod_format_DT_ui(id)

mod_format_DT_server(
  id,
  data = reactive({
     NULL
 }),
  withDLBtns = FALSE,
  showRownames = FALSE,
  dom = "Bt",
  hc_style = reactive({
     NULL
 }),
  full_style = reactive({
     NULL
 }),
  filename = "Prostar_export",
  hideCols = reactive({
     NULL
 }),
  selection = "single"
)
}
\arguments{
\item{id}{shiny id}

\item{data}{xxx}

\item{withDLBtns}{xxx}

\item{showRownames}{xxx}

\item{dom}{xxx}

\item{hc_style}{xxx}

\item{full_style}{A list of four items:
\itemize{
\item cols: a vector of colnames of columns to show,
\item vals: a vector of colnames of columns that contain values,
\item unique: unique(conds),
\item pal: xxxxx
}}

\item{filename}{xxx}

\item{hideCols}{xxx}

\item{selection}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
