% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_format_DT.R
\name{mod_format_DT}
\alias{mod_format_DT}
\alias{mod_format_DT_ui}
\alias{mod_format_DT_server}
\title{mod_format_DT_ui and mod_format_DT_server}
\usage{
mod_format_DT_ui(id)

mod_format_DT_server(
  id,
  data = reactive({
     NULL
 }),
  data_nostyle = reactive({
     NULL
 }),
  withDLBtns = FALSE,
  showRownames = FALSE,
  dom = "Bt",
  dt_style = reactive({
     NULL
 }),
  filename = "Prostar_export",
  hideCols = reactive({
     NULL
 }),
  selection = "single"
)
}
\arguments{
\item{id}{shiny id}

\item{data}{A \code{data.frame}}

\item{withDLBtns}{A boolean to indicate whether to display download buttons or not.}

\item{showRownames}{A boolean to indicate whether to show rownames.}

\item{dom}{xxx}

\item{dt_style}{A \code{list} composed of:
\itemize{
\item data : a data.frame` xxxx
\item colors : a named vector
}}

\item{filename}{A \code{character(1)} which is the default filename for download.}

\item{hideCols}{xxx}

\item{selection}{xxx}

\item{dataForStyling}{xxx}

\item{xls_style}{A list of four items:
\itemize{
\item cols: a vector of colnames of columns to show,
\item vals: a vector of colnames of columns that contain values,
\item unique: unique(conds),
\item pal: RColorBrewer::brewer.pal(3, "Dark2")\link{1:2}
}}
}
\value{
NA
}
\description{
A shiny Module.

See \code{DT} package homepage for more details about styling tables.
If no style is precised, this module show the raw data.
If any style is given, then the dataset must be well configured (I.e. it
must contain the correct columns )
}
