% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_tracking.R
\name{mod_plots_tracking}
\alias{mod_plots_tracking}
\alias{mod_plots_tracking_ui}
\alias{mod_plots_tracking_server}
\title{mod_plots_tracking_ui and mod_plots_tracking_server}
\usage{
mod_plots_tracking_ui(id)

mod_plots_tracking_server(
  id,
  vizData = reactive({
     NULL
 }),
  resetBtn = reactive({
     FALSE
 })
)
}
\arguments{
\item{id}{shiny id}

\item{vizData}{internal}

\item{resetBtn}{A \code{boolean(1)} which indicates whether to show the 'Reset' button
or not.}

\item{params}{A \code{list} of three items to give instructions to the module
in this it is run in slave mode:
\itemize{
\item type: xxxxx
\item names: xxxxx
\item indices: xxxxx
}}
}
\value{
A \code{list} (same structure as the parameter 'params")
}
\description{
This shiny module provides a tool to select
}
\examples{
library(DaparViz)
library(highcharter)
library(MSnbase)
library(shiny)

options(shiny.fullstacktrace = TRUE)



ui <- fluidPage(
  
  mod_plots_tracking_ui("tracker")
)



server <- function(input, output, session) {
  # data(ft, package='DaparToolshed')
  # data(ft_na, package='DaparToolshed')
  # vList <- convert2viz(ft)
  # vList_na <- convert2viz(ft_na)
  
  #
  # Example with a QFeatures dataset
  #
  data(Exp1_R25_pept, package='DaparToolshedData')
  vList <- convert2viz(Exp1_R25_pept)
  vData <- vList@ll.vizData[[1]]
  
  #
  # Example with a series of MSnSet datasets
  #
  # data(Exp1_R25_prot, package='DAPARdata')
  # data(Exp1_R25_pept, package='DAPARdata')
  # data(Exp1_R2_pept, package='DAPARdata')
  # ll.tmp <- setNames(c(Exp1_R25_prot, Exp1_R25_pept, Exp1_R2_pept),
  #                    nm = c('Exp1_R25_prot', 'Exp1_R25_pept', 'Exp1_R2_pept'))
  # 
  # vList <- convert2viz(ll.tmp)
  
  
  
  indices <- mod_plots_tracking_server("tracker", vizData = reactive({vData}))
  
  observe({
    print(indices())
    
  })
}

if (interactive())
  shinyApp(ui, server)
}
\keyword{internal}
