% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_tracking.R
\name{mod_plots_tracking_ui}
\alias{mod_plots_tracking_ui}
\alias{mod_plots_tracking_server}
\title{mod_plots_boxplots_ui and mod_plots_boxplots_server}
\usage{
mod_plots_tracking_ui(id)

mod_plots_tracking_server(
  input,
  output,
  session,
  obj,
  keyId,
  params,
  reset = FALSE,
  slave = FALSE
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{metadata}{Metadata of Features con}
}
\description{
A shiny Module.
}
\keyword{internal}
