% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_tabExplorer.R
\name{omXplore_tabExplorer}
\alias{omXplore_tabExplorer}
\alias{omXplore_tabExplorer_ui}
\alias{omXplore_tabExplorer_server}
\title{Explore \code{VizData} objects.}
\usage{
omXplore_tabExplorer_ui(id)

omXplore_tabExplorer_server(
  id,
  obj = reactive({
     NULL
 }),
  digits = reactive({
     3
 })
)

omXplore_tabExplorer(obj)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{An instance of the class \code{VizData}}

\item{digits}{xxx}
}
\value{
NA

NA

NA

A shiny app
}
\description{
xxx
}
\examples{
if (interactive()) {
  data(vData_ft)
  omXplore_tabExplorer(vData_ft[[1]])
}

}
