% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_CC.R
\name{pep_prot_CC}
\alias{pep_prot_CC}
\alias{buildGraph}
\alias{display.CC.visNet}
\alias{plotCCJitter}
\title{Display a CC}
\usage{
buildGraph(cc, metadata = NULL)

display.CC.visNet(
  g,
  layout = "layout_with_fr",
  obj = NULL,
  prot.tooltip = NULL,
  pept.tooltip = NULL
)

plotCCJitter(df, clickFunction = NULL)
}
\arguments{
\item{cc}{A cc (a list)}

\item{metadata}{xxx}

\item{g}{A cc (a list)}

\item{layout}{xxxxx}

\item{obj}{xxx}

\item{prot.tooltip}{xxx}

\item{pept.tooltip}{xxx}

\item{df}{xxxx}

\item{clickFunction}{xxxx}
}
\value{
A list

A plot

A plot
}
\description{
Display a CC
}
\examples{
data(vdata)
g <- buildGraph(get_cc(vdata[[1]]))
display.CC.visNet(g)

n.prot <- unlist(lapply(GetSlotCc(vdata[[1]]), function(x) {ncol(x)}))
n.pept <- unlist(lapply(GetSlotCc(vdata[[1]]), function(x) {nrow(x)}))
df <- tibble::tibble(
  x = jitter(n.pept),
  y = jitter(n.prot),
  index = seq_len(length(GetSlotCc(vdata[[1]])))
)
plotCCJitter(df)

}
\author{
Thomas Burger, Samuel Wieczorek
}
