% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_CC.R
\name{pep_prot_CC}
\alias{pep_prot_CC}
\alias{buildGraph}
\alias{display.CC.visNet}
\alias{plotCCJitter}
\title{Display a CC}
\usage{
buildGraph(cc, metadata = NULL)

display.CC.visNet(
  g,
  layout = "layout_with_fr",
  obj = NULL,
  prot.tooltip = NULL,
  pept.tooltip = NULL
)

plotCCJitter(df, clickFunction = NULL)
}
\arguments{
\item{cc}{A cc (a list)}

\item{metadata}{xxx}

\item{g}{A cc (a list)}

\item{layout}{xxxxx}

\item{obj}{xxx}

\item{prot.tooltip}{xxx}

\item{pept.tooltip}{xxx}

\item{df}{xxxx}

\item{clickFunction}{xxxx}
}
\value{
A list

A plot

A plot
}
\description{
Display a CC
}
\examples{
data(vData_ms)
obj <- vData_ms[[1]]
g <- buildGraph(GetSlotCc(obj)[[1]])


data(vData_ft)
cc <- GetSlotCc(vData_ft[[1]])[[1]]
g <- buildGraph(cc)
display.CC.visNet(g)

data(vData_ft)
obj <- vData_ft[[1]]

n.prot <- unlist(lapply(GetSlotCc(obj), function(x) {
  ncol(x)
}))
n.pept <- unlist(lapply(GetSlotCc(obj), function(x) {
  nrow(x)
}))
df <- tibble::tibble(
  x = jitter(n.pept),
  y = jitter(n.prot),
  index = seq_len(length(GetSlotCc(obj)))
)
plotCCJitter(df)

}
\author{
Thomas Burger, Samuel Wieczorek
}
