% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mv_imputation.R, R/plots_mv_for_impute.R
\name{plot-mv}
\alias{plot-mv}
\alias{mod_mv_imputation_ui}
\alias{mod_mv_imputation_server}
\alias{mv.density}
\alias{mv.mec.heatmap}
\title{Distribution of Observed values with respect to intensity values}
\usage{
mod_mv_imputation_ui(id)

mod_mv_imputation_server(
  id,
  se = reactive({
     NULL
 }),
  conds = NULL,
  pal.name = reactive({
     NULL
 })
)

mv.density(se, conds, pal.name = NULL, pattern, typeofMV = NULL, title = NULL)

mv.mec.heatmap(se, conds)
}
\arguments{
\item{id}{A \code{character(1)} which is the 'id' of the shiny module.}

\item{se}{An instance of a class \code{SummarizedExperiment}.}

\item{conds}{A \code{character()} of the name of conditions (one condition
per sample).}

\item{pal.name}{A \code{character(1)} which is the name of the palette (from
the package \link{RColorBrewer} to use.}

\item{pattern}{xxx}

\item{typeofMV}{xxx}

\item{title}{The title of the plot}
}
\value{
A plot
}
\description{
This method shows density plots which represents the repartition of
Partial Observed Values for each replicate in the dataset.
The colors correspond to the different conditions (Condition in colData of
the object of class \code{QFeatures}).
The x-axis represent the mean of intensity for one condition and one
entity in the dataset (i. e. a protein)
whereas the y-axis count the number of observed values for this entity
and the considered condition.
}
\section{xxx}{


xxxx
}

\section{missing value image}{

Plots a heatmap of the quantitative data. Each column represent one of
the conditions in the object of class \code{MSnSet} and
the color is proportional to the mean of intensity for each line of
the dataset.
The lines have been sorted in order to visualize easily the different
number of missing values. A white square is plotted for missing values.
}

\examples{
library(SummarizedExperiment)
data(ft_na, package='DaparViz')
data <- assay(ft_na, 1)
conds <- colData(ft_na)$Condition

#-----------------------------
# xxx
#-----------------------------

mv.density(ft_na[[1]], conds, pattern = "missing MEC")

mv.mec.heatmap(ft_na[[1]], conds)

#-----------------------------
# xxx
#-----------------------------

if (interactive()) {
    data(ft_na, package='DaparViz')

    ui <- mod_mv_imputation_ui("plot")

    server <- function(input, output, session) {
        mod_mv_imputation_server("plot",
            se = reactive({
                ft_na[[1]]
            }),
            conds = colData(ft_na)$Condition
        )
    }

    shinyApp(ui = ui, server = server)
}
}
\seealso{
The package \code{DaparToolshed} for the 'qMetadata' metadata.
}
\author{
Samuel Wieczorek, Enora Fremy
}
