% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_variance.R, R/plot_variance.R
\name{plot-variance}
\alias{plot-variance}
\alias{mod_ds_variance_ui}
\alias{mod_ds_variance_server}
\alias{CVDist}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_variance_ui(id)

mod_ds_variance_server(id, data, conds, pal.name = NULL)

CVDist(data, conds = NULL, pal.name = NULL)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{data}{A \code{data.frame} or \code{matrix} which is the quantitative values.}

\item{conds}{A \code{character()} of the name of conditions
(one condition per sample).}

\item{pal.name}{A \code{character(1)} which is the name of the palette from the
package \link{RColorBrewer} from which the colors are taken.
Default value is 'Set1'.}
}
\value{
NA

NA
}
\description{
xxxx
}
\examples{
library(SummarizedExperiment)
data(ft, package='DaparViz')
CVDist(assay(ft, 1), colData(ft)$Condition)


#------------------------------------------
# Shiny module
#------------------------------------------
if (interactive()) {
    data(ft, package='DaparViz')
    ui <- mod_ds_variance_ui("plot")

    server <- function(input, output, session) {
        mod_ds_variance_server(
            "plot",
            reactive({
                assay(ft, 1)
            })
        )
    }

    shinyApp(ui = ui, server = server)
}

}
