% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_variance.R, R/plot_variance.R
\name{plot-variance}
\alias{plot-variance}
\alias{mod_ds_variance_ui}
\alias{mod_ds_variance_server}
\alias{ds_variance}
\alias{CVDist}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_variance_ui(id)

mod_ds_variance_server(id, obj, pal.name = NULL)

ds_variance(obj)

CVDist(obj, pal.name = NULL)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{xxx}

\item{pal.name}{xxx}
}
\value{
NA

NA
}
\description{
xxxx
}
\examples{
data(vData_ft)
ds_variance(vDataz_ft[[1]])

}
