% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_variance.R, R/plot_variance.R
\name{plot-variance}
\alias{plot-variance}
\alias{mod_ds_variance_ui}
\alias{mod_ds_variance_server}
\alias{CVDist}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_variance_ui(id)

mod_ds_variance_server(id, vizData, pal.name = NULL)

CVDist(vizData, pal.name = NULL)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{vizData}{xxx}

\item{pal.name}{xxx}
}
\value{
NA

NA
}
\description{
xxxx
}
\examples{
library(DaparViz)


ui <- fluidPage(mod_ds_variance_ui("plot"))

server <- shinyServer(function(input, output, session) {
  data(vData_ft)
  obj <- vData_ft[1]
  
  mod_ds_variance_server("plot", 
                         vizData = reactive({obj}))
})

if (interactive())
  shinyApp(ui, server)
library(DaparViz)


ui <- fluidPage(mod_ds_variance_ui("plot"))

server <- shinyServer(function(input, output, session) {
  data(vData_ft)
  obj <- vData_ft[1]
  
  mod_ds_variance_server("plot", 
                         vizData = reactive({obj}))
})

if (interactive())
  shinyApp(ui, server)
}
