% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_variance.R, R/plot_variance.R
\name{plot-variance}
\alias{plot-variance}
\alias{mod_ds_variance_ui}
\alias{mod_ds_variance_server}
\alias{CVDist}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_variance_ui(id)

mod_ds_variance_server(id, vData, pal.name = NULL)

CVDist(vizData, pal.name = NULL)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{vData}{xxx}

\item{pal.name}{A \code{character(1)} which is the name of the palette from the
package \link{RColorBrewer} from which the colors are taken.
Default value is 'Set1'.}
}
\value{
NA

NA
}
\description{
xxxx
}
\examples{
library(highcharter)
library(DaparViz)
data(ft, package='DaparToolshed')
vList <- convert2viz(ft)
vData <- vList@ll.vizData[[1]]


CVDist(vData)


#------------------------------------------
# Shiny module
#------------------------------------------

ui <- shinyUI(mod_ds_variance_ui("plot"))

server <- shinyServer(function(input, output, session) {
  data(ft, package='DaparToolshed')
  vList <- convert2viz(ft)
  vData <- vList@ll.vizData[[1]]
  
  mod_ds_variance_server("plot", vData = reactive({vData}))
})

shinyApp(ui, server)
}
