% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_variance.R, R/plot_variance.R
\name{plot-variance}
\alias{plot-variance}
\alias{mod_ds_variance_ui}
\alias{mod_ds_variance_server}
\alias{CVDist}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_variance_ui(id)

mod_ds_variance_server(id, vizData, pal.name = NULL)

CVDist(vizData, pal.name = NULL)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{vizData}{xxx}

\item{pal.name}{xxx}
}
\value{
NA

NA
}
\description{
xxxx
}
\examples{
library(highcharter)
library(DaparViz)
library(shiny)


data(ft)
vList <- convert2viz(ft)
vData <- vList@ll.vizData[[1]]


CVDist(vData)


#------------------------------------------
# Shiny module
#------------------------------------------

ui <- fluidPage(mod_ds_variance_ui("plot"))

server <- shinyServer(function(input, output, session) {
  data(ft)
  vList <- convert2viz(ft)
  vData <- vList@ll.vizData[[1]]
  
  #vData <- ft
  
  mod_ds_variance_server("plot", vizData = reactive({vData}))
})

if (interactive())
  shinyApp(ui, server)
}
