% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparViz_variance.R
\name{plot-variance}
\alias{plot-variance}
\alias{DaparViz_variance_ui}
\alias{DaparViz_variance_server}
\alias{CVDist}
\alias{DaparViz_variance}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
DaparViz_variance_ui(id)

DaparViz_variance_server(id, obj, pal.name = NULL)

CVDist(obj, pal.name = NULL)

DaparViz_variance(obj)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{obj}{xxx}

\item{pal.name}{xxx}
}
\value{
NA

NA

A plot

A shiny app
}
\description{
xxxx
}
\examples{
if (interactive()) {
  data(vData_ft)
  DaparViz_variance(vData_ft[[1]])
}

}
