% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_variance.R, R/plot_variance.R
\name{plot-variance}
\alias{plot-variance}
\alias{mod_ds_variance_ui}
\alias{mod_ds_variance_server}
\alias{CVDist}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
mod_ds_variance_ui(id)

mod_ds_variance_server(id, data, conds, pal.name = NULL)

CVDist(data, conds = NULL, pal.name = NULL)
}
\arguments{
\item{id}{shiny id}

\item{data}{xxx}

\item{conds}{A \code{character()} representings the condition for
each sample of the \link{QFeatures} object.}

\item{pal.name}{xxx}
}
\description{
xxxx
}
\examples{
library(QFeatures)
data(ft)
CVDist(assay(ft, 1), colData(ft)$Condition)


#------------------------------------------
# Shiny module
#------------------------------------------
if(interactive()){
data(ft)
 ui <- mod_ds_variance_ui('plot')

 server <- function(input, output, session) {
  mod_ds_variance_server('plot', 
  reactive({assay(ft, 1)})
  )}
 
 shinyApp(ui=ui, server=server)
}




}
