% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omXplore_plots_tracking.R
\name{plots_tracking}
\alias{plots_tracking}
\alias{plots_tracking_ui}
\alias{plots_tracking_server}
\title{plots_tracking_ui and plots_tracking_server}
\usage{
plots_tracking_ui(id)

plots_tracking_server(
  id,
  dataIn = reactive({
     NULL
 }),
  remoteReset = reactive({
     NULL
 }),
  is.enabled = reactive({
     TRUE
 })
)

plots_tracking(obj)
}
\arguments{
\item{id}{shiny id}

\item{remoteReset}{A \code{boolean(1)} which indicates whether to show the 'Reset'
button or not.}

\item{is.enabled}{xxx}

\item{obj}{An instance of the class \code{MultiAssayExperiment}}
}
\value{
NA

A \code{list} (same structure as the parameter \code{params})

A shiny app
}
\description{
This shiny module provides a tool to select
}
\examples{
\dontrun{
  data(vdata)
  shiny::runApp(plots_tracking(vdata[[1]]))
  }

}
\keyword{internal}
