% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparViz_plots_tracking.R
\name{plots_tracking}
\alias{plots_tracking}
\alias{plots_tracking_ui}
\alias{plots_tracking_server}
\title{plots_tracking_ui and plots_tracking_server}
\usage{
plots_tracking_ui(id)

plots_tracking_server(
  id,
  obj = reactive({
     NULL
 }),
  resetBtn = reactive({
     FALSE
 })
)
}
\arguments{
\item{id}{shiny id}

\item{obj}{An instance of the class \code{DaparViz}}

\item{resetBtn}{A \code{boolean(1)} which indicates whether to show the 'Reset' button
or not.}

\item{params}{A \code{list} of three items to give instructions to the module
in this it is run in slave mode:
\itemize{
\item type: xxxxx
\item names: xxxxx
\item indices: xxxxx
}}
}
\value{
A \code{list} (same structure as the parameter \code{params})
}
\description{
This shiny module provides a tool to select
}
\keyword{internal}
