% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell_utils.R
\name{q_metadata}
\alias{q_metadata}
\alias{metacell.def}
\title{Quantitative metadata vocabulary for entities}
\usage{
metacell.def(level)
}
\arguments{
\item{level}{A string designing the type of entity/pipeline.
Available values are: \code{peptide}, \code{protein}}
}
\value{
A data.frame containing the different tags and corresponding colors for the level
given in parameter
}
\description{
This function gives the vocabulary used for the quantitative metadata of each entity in
each condition.
}
\section{Glossary}{


Peptide-level vocabulary

|-- 'Any'
|    |
|    |-- 1.0 'Quantified'
|    |    |
|    |    |-- 1.1 "Quant. by direct id" (color 4, white)
|    |    |
|    |    |-- 1.2 "Quant. by recovery" (color 3, lightgrey)
|    |
|    |-- 2.0 "Missing" (no color)
|    |    |
|    |    |-- 2.1 "Missing POV" (color 1)
|    |    |
|    |    |-- 2.2 'Missing MEC' (color 2)
|    |
|    |-- 3.0 'Imputed'
|    |    |
|    |    |-- 3.1 'Imputed POV' (color 1)
|    |    |
|    |    |-- 3.2 'Imputed MEC' (color 2)

Protein-level vocabulary:
|-- 'Any'
|    |
|    |-- 1.0 'Quantified'
|    |    |
|    |    |-- 1.1 "Quant. by direct id" (color 4, white)
|    |    |
|    |    |-- 1.2 "Quant. by recovery" (color 3, lightgrey)
|    |
|    |-- 2.0 "Missing"
|    |    |
|    |    |-- 2.1 "Missing POV" (color 1)
|    |    |
|    |    |-- 2.2 'Missing MEC' (color 2)
|    |
|    |-- 3.0 'Imputed'
|    |    |
|    |    |-- 3.1 'Imputed POV' (color 1)
|    |    |
|    |    |-- 3.2 'Imputed MEC' (color 2)
|    |
|    |-- 4.0 'Combined tags' (color 3bis, lightgrey)
}

\section{Conversion to the glossary}{


A generic conversion

Conversion for Proline datasets

Conversion from Maxquant datasets
}

\examples{

metacell.def('protein')
metacell.def('peptide')

#-----------------------------------------------
# A shiny app to view color legends 
#-----------------------------------------------
if(interactive()){
data(ft)
ui <- mod_qMetacellLegend_ui("legend")

server <- function(input, output, session) {
  mod_qMetacellLegend_server('legend',
                                object = reactive({ft[[1]]}))
  }
  
 shinyApp(ui = ui, server = server)
  
  
}
}
\author{
Thomas Burger, Samuel Wieczorek
}
