% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_seExplorer.R
\name{SE-explorer}
\alias{SE-explorer}
\alias{mod_ds_seExplorer_ui}
\alias{mod_ds_seExplorer_server}
\title{Explore a \code{SummarizedExperiment} object}
\usage{
mod_ds_seExplorer_ui(id)

mod_ds_seExplorer_server(
  id,
  vData,
  digits = reactive({
     3
 })
)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{digits}{xxx}

\item{se}{An instance of the class \code{SummarizedExperiment}}
}
\value{
NA
}
\description{
xxxx
}
\examples{
library(MSnbase)
library(DaparToolshed)
library(highcharter)
library(DaparViz)


data(ft, package='DaparToolshed')
vList <- convert2viz(ft)
vData <- vList@ll.vizData[[1]]

ui <- mod_ds_seExplorer_ui("plot")

server <- function(input, output, session) {
        mod_ds_seExplorer_server("plot", reactive({vData})
        )
    }

shinyApp(ui = ui, server = server)
}
