% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_seExplorer.R
\name{SE-explorer}
\alias{SE-explorer}
\alias{mod_ds_seExplorer_ui}
\title{Explore a \code{SummarizedExperiment} object}
\usage{
mod_ds_seExplorer_ui(id)
}
\arguments{
\item{id}{xxx}
}
\description{
xxxx
}
\examples{
library(QFeatures)
data(ft)
corrMatrix(assay(ft, 1))


#------------------------------------------
# Shiny module
#------------------------------------------
if(interactive()){
library(DaparToolshed)
data(ft_na)
 ui <- mod_ds_seExplorer_ui('plot')

 server <- function(input, output, session) {
  mod_ds_seExplorer_server('plot', 
  reactive({ft_na[[1]]})
  )}
 
 shinyApp(ui=ui, server=server)
}
}
