% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_seExplorer.R
\name{SE-explorer}
\alias{SE-explorer}
\alias{mod_ds_seExplorer_ui}
\alias{mod_ds_seExplorer_server}
\title{Explore a \code{VizClass} object}
\usage{
mod_ds_seExplorer_ui(id)

mod_ds_seExplorer_server(
  id,
  vizData = reactive({
     NULL
 }),
  digits = reactive({
     3
 })
)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{digits}{xxx}

\item{se}{An instance of the class \code{VizData}}
}
\value{
NA
}
\description{
xxxx
}
