% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_seExplorer.R
\name{SE-explorer}
\alias{SE-explorer}
\alias{mod_ds_seExplorer_ui}
\alias{mod_ds_seExplorer_server}
\title{Explore a \code{VizClass} object}
\usage{
mod_ds_seExplorer_ui(id)

mod_ds_seExplorer_server(
  id,
  vizData = reactive({
     NULL
 }),
  digits = reactive({
     3
 })
)
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{vizData}{An instance of the class \code{VizData}}

\item{digits}{xxx}
}
\value{
NA
}
\description{
xxx
}
\examples{
library(MSnbase)
#library(DaparToolshed)
library(highcharter)
library(DaparViz)
library(shiny)



ui <- fluidPage(
  mod_ds_seExplorer_ui("plot")
)

server <- function(input, output, session) {
  data(ft)
  vList <- convert2viz(ft)
  vData <- vList@ll.vizData[[1]]
  
  mod_ds_seExplorer_server("plot", reactive({vData}))
  
    }

if (interactive())
  shinyApp(ui = ui, server = server)
}
