% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ds_seExplorer.R
\name{SE-explorer}
\alias{SE-explorer}
\alias{mod_ds_seExplorer_ui}
\alias{mod_ds_seExplorer_server}
\title{Explore a \code{SummarizedExperiment} object}
\usage{
mod_ds_seExplorer_ui(id)

mod_ds_seExplorer_server(id, se, digits = reactive({
    3
}))
}
\arguments{
\item{id}{A \code{character(1)} which is the id of the shiny module.}

\item{se}{An instance of the class \code{SummarizedExperiment}}

\item{digits}{xxx}
}
\value{
NA
}
\description{
xxxx
}
\examples{
data(ft)
corrMatrix(assay(ft, 1))


#------------------------------------------
# Shiny module
#------------------------------------------
if(interactive()){
data(ft_na)
 ui <- mod_ds_seExplorer_ui('plot')

 server <- function(input, output, session) {
  mod_ds_seExplorer_server('plot', 
  reactive({ft_na[[1]]})
  )}
 
 shinyApp(ui=ui, server=server)
}
}
