% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_seTracker.R
\name{tracking}
\alias{tracking}
\alias{mod_seTracker_ui}
\alias{mod_seTracker_server}
\title{Tracking of entities within plots}
\usage{
mod_seTracker_ui(id)

mod_seTracker_server(id, se)
}
\arguments{
\item{id}{xxx}

\item{se}{A instance of the class \code{SummarizedExperiment}}
}
\description{
This shiny module offers a UI to select a subset of a dataset
and superimpose quantitative values of this selection on the complete plot
Three modes of selection are implemented:
\itemize{
\item 'Protein list': xxx,
\item 'Random': xxx,
\item 'Specific column': xxx
}
}
\examples{
if (interactive()){

ui <- tagList(
  mod_seTracker_ui('track'),
  uiOutput('show')
)

server <- function(input, output, session) {
  tmp <- reactiveVal()
  
  tmp <- mod_seTracker_server(id = 'track',
                              se = reactive({ft[[1]]})
  )
  
  output$show <- renderUI({
    p(paste0(tmp(), collapse = ' '))
  })
  
}
shinyApp(ui=ui, server=server)
}
}
