% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizData_Class.R
\docType{class}
\name{VizData-class}
\alias{VizData-class}
\alias{VizData}
\alias{show,VizData-method}
\alias{initialize,VizData-method}
\title{VizData class}
\usage{
\S4method{show}{VizData}(object)

\S4method{initialize}{VizData}(
  .Object,
  qdata,
  metacell,
  metadata,
  colID,
  proteinID,
  conds,
  type,
  adjMat,
  cc
)
}
\arguments{
\item{object}{xxx}
}
\description{
Conceptually, a \code{VizData} object is a standard representation of all
elements from quantitative other structured data used in proteomics, such
as \code{MSnset} or \code{QFeatures}. It allows to use it as a generic converter.

\code{VizData} objects are not usually created by hand but are created from the \code{VizList}
which is the class accessible by the user.

The recommended way to create \code{VizData} objects is the use the
\code{xxxx()} function of the class \code{VizList}
}
\section{Slots}{

\describe{
\item{\code{qdata}}{xxx}

\item{\code{metacell}}{xxx}

\item{\code{metadata}}{xxx}

\item{\code{colID}}{xxx}

\item{\code{proteinID}}{xxx}

\item{\code{conds}}{xxx}

\item{\code{type}}{xxx}

\item{\code{adjMat}}{xxx}

\item{\code{cc}}{xxx}
}}

