% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizData_Class.R
\docType{class}
\name{VizData-class}
\alias{VizData-class}
\alias{VizData}
\alias{show,VizData-method}
\alias{initialize,VizData-method}
\title{VizData class}
\usage{
\S4method{show}{VizData}(object)

\S4method{initialize}{VizData}(
  .Object,
  qdata,
  metacell,
  metadata,
  colID,
  proteinID,
  conds,
  type,
  adjMat,
  cc
)
}
\arguments{
\item{object}{xxx}
}
\description{
Conceptually, a \code{VizData} object is a standard representation of all
elements from quantitative other structured data used in proteomics, such
as \code{MSnset} or \code{QFeatures}. It allows to use it as a generic converter.

\code{VizData} objects are not usually created by hand but are created from the \code{VizList}
which is the class accessible by the user.

The recommended way to create \code{VizData} objects is the use the
\code{xxxx()} function of the class \code{VizList}
}
\section{Slots}{

\describe{
\item{\code{qdata}}{xxx}

\item{\code{metacell}}{xxx}

\item{\code{metadata}}{xxx}

\item{\code{colID}}{xxx}

\item{\code{proteinID}}{xxx}

\item{\code{conds}}{xxx}

\item{\code{type}}{xxx}

\item{\code{adjMat}}{xxx}

\item{\code{cc}}{xxx}
}}

\examples{
library(MSnbase)
#library(DaparToolshed)
library(DaparViz)
library(Matrix)

# Convert a QFeatures object to an instance of VizList class
data(Exp1_R25_pept, package='DaparToolshedData')
test <- convert2viz(Exp1_R25_pept)

# Convert a list of MSnSet objects to an instance of VizList class
data(Exp1_R25_prot, package='DAPARdata')
data(Exp1_R25_pept, package='DAPARdata')
data(Exp1_R2_pept, package='DAPARdata')
ll.tmp <- setNames(c(Exp1_R25_prot, Exp1_R25_pept, Exp1_R2_pept),
                   nm = c('Exp1_R25_prot', 'Exp1_R25_pept', 'Exp1_R2_pept'))

test <- convert2viz(ll.tmp)



data(Exp1_R25_pept, package='DAPARdata')
msnset <- Exp1_R25_pept
X <- PSMatch::makeAdjacencyMatrix(MSnbase::fData(msnset)[, msnset@experimentData@other$proteinId])
rownames(X) <- rownames(MSnbase::fData(msnset))
connectedComp <- PSMatch::ConnectedComponents(X)

ll.Exp1_R25_pept <- list(
  qdata = MSnbase::exprs(msnset),
  metacell = MSnbase::fData(msnset)[ , msnset@experimentData@other$names_metacell],
  mdata = MSnbase::fData(msnset),
  colID = msnset@experimentData@other$keyId,
  conds = MSnbase::pData(msnset)[, 'Condition'],
  type = msnset@experimentData@other$typeOfData,
  adjMat = as.matrix(X),
  cc = as.list(connectedComp@adjMatrices)
  )

data(Exp1_R25_prot, package='DAPARdata')
msnset <- Exp1_R25_prot
ll.Exp1_R25_prot <- list(
   qdata = MSnbase::exprs(msnset),
  metacell = MSnbase::fData(msnset)[ , msnset@experimentData@other$names_metacell],
  mdata = MSnbase::fData(msnset),
  colID = msnset@experimentData@other$keyId,
  conds = MSnbase::pData(msnset)[, 'Condition'],
  type = msnset@experimentData@other$typeOfData
)

data(Exp1_R2_pept, package='DAPARdata')
msnset <- Exp1_R2_pept
X <- PSMatch::makeAdjacencyMatrix(MSnbase::fData(msnset)[, msnset@experimentData@other$proteinId])
rownames(X) <- rownames(MSnbase::fData(msnset))
connectedComp <- PSMatch::ConnectedComponents(X)

ll.Exp1_R2_pept <- list(
  qdata = MSnbase::exprs(msnset),
  metacell = MSnbase::fData(msnset)[ , msnset@experimentData@other$names_metacell],
  mdata = MSnbase::fData(msnset),
  colID = msnset@experimentData@other$keyId,
  conds = MSnbase::pData(msnset)[, 'Condition'],
  type = msnset@experimentData@other$typeOfData,
  adjMat = as.matrix(X),
  cc = as.list(connectedComp@adjMatrices)
  
)


ll.tmp <- list(
  Exp1_R25_pept = ll.Exp1_R25_pept,
  Exp1_R25_prot = ll.Exp1_R25_prot,
  Exp1_R2_pept = ll.Exp1_R2_pept
)

convert2viz(ll.tmp)
}
