% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizList_Class.R
\name{VizList-class}
\alias{VizList-class}
\alias{VizList}
\alias{initialize,VizList-method}
\alias{GetVizData}
\title{DaparVizData class definition}
\usage{
VizList(ll.vizData = NULL)

\S4method{initialize}{VizList}(.Object, ll.vizData)

VizList(ll.vizData = NULL)

GetVizData(object, i)
}
\arguments{
\item{ll.vizData}{xxx}

\item{.Object}{xxx}

\item{object}{An instance of the class \code{VizList}}

\item{i}{An \code{integer(1)}}
}
\value{
An instance of 'VizList' class

An instance of the class \code{VizData}
}
\description{
xxx

xxx
}
\section{Slots}{

\describe{
\item{\code{ll.vizData}}{xxx}
}}

