% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizList_Class.R
\name{VizList-class}
\alias{VizList-class}
\alias{VizList}
\alias{initialize,VizList-method}
\alias{GetVizData}
\title{DaparVizData class definition}
\usage{
VizList(ll.vizData = NULL)

\S4method{initialize}{VizList}(.Object, ll.vizData)

VizList(ll.vizData = NULL)

GetVizData(object, i)
}
\arguments{
\item{ll.vizData}{xxx}

\item{.Object}{xxx}

\item{object}{An instance of the class \code{VizList}}

\item{i}{An \code{integer(1)}}
}
\value{
An instance of 'VizList' class

An instance of the class \code{VizData}
}
\description{
xxx

xxx
}
\section{Slots}{

\describe{
\item{\code{ll.vizData}}{xxx}
}}

\examples{
library(MSnbase)
#library(DaparToolshed)
library(DaparViz)
library(Matrix)

# Convert a QFeatures object to an instance of VizList class
data(Exp1_R25_pept, package='DaparToolshedData')
test <- convert2viz(Exp1_R25_pept)

# Convert a list of MSnSet objects to an instance of VizList class
data(Exp1_R25_prot, package='DAPARdata')
data(Exp1_R25_pept, package='DAPARdata')
data(Exp1_R2_pept, package='DAPARdata')
ll.tmp <- setNames(c(Exp1_R25_prot, Exp1_R25_pept, Exp1_R2_pept),
                   nm = c('Exp1_R25_prot', 'Exp1_R25_pept', 'Exp1_R2_pept'))

test <- convert2viz(ll.tmp)



data(Exp1_R25_pept, package='DAPARdata')
msnset <- Exp1_R25_pept
X <- PSMatch::makeAdjacencyMatrix(MSnbase::fData(msnset)[, msnset@experimentData@other$proteinId])
rownames(X) <- rownames(MSnbase::fData(msnset))
connectedComp <- PSMatch::ConnectedComponents(X)

ll.Exp1_R25_pept <- list(
  qdata = MSnbase::exprs(msnset),
  metacell = MSnbase::fData(msnset)[ , msnset@experimentData@other$names_metacell],
  mdata = MSnbase::fData(msnset),
  colID = msnset@experimentData@other$keyId,
  conds = MSnbase::pData(msnset)[, 'Condition'],
  type = msnset@experimentData@other$typeOfData,
  adjMat = as.matrix(X),
  cc = as.list(connectedComp@adjMatrices)
  )

data(Exp1_R25_prot, package='DAPARdata')
msnset <- Exp1_R25_prot
ll.Exp1_R25_prot <- list(
   qdata = MSnbase::exprs(msnset),
  metacell = MSnbase::fData(msnset)[ , msnset@experimentData@other$names_metacell],
  mdata = MSnbase::fData(msnset),
  colID = msnset@experimentData@other$keyId,
  conds = MSnbase::pData(msnset)[, 'Condition'],
  type = msnset@experimentData@other$typeOfData
)

data(Exp1_R2_pept, package='DAPARdata')
msnset <- Exp1_R2_pept
X <- PSMatch::makeAdjacencyMatrix(MSnbase::fData(msnset)[, msnset@experimentData@other$proteinId])
rownames(X) <- rownames(MSnbase::fData(msnset))
connectedComp <- PSMatch::ConnectedComponents(X)

ll.Exp1_R2_pept <- list(
  qdata = MSnbase::exprs(msnset),
  metacell = MSnbase::fData(msnset)[ , msnset@experimentData@other$names_metacell],
  mdata = MSnbase::fData(msnset),
  colID = msnset@experimentData@other$keyId,
  conds = MSnbase::pData(msnset)[, 'Condition'],
  type = msnset@experimentData@other$typeOfData,
  adjMat = as.matrix(X),
  cc = as.list(connectedComp@adjMatrices)
  
)


ll.tmp <- list(
  Exp1_R25_pept = ll.Exp1_R25_pept,
  Exp1_R25_prot = ll.Exp1_R25_prot,
  Exp1_R2_pept = ll.Exp1_R2_pept
)

convert2viz(ll.tmp)
}
