% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizData_Class.R, R/VizList_Class.R
\name{Convert2VizList,QFeatures-method}
\alias{Convert2VizList,QFeatures-method}
\alias{VizList-class}
\alias{VizList}
\alias{initialize,VizList-method}
\alias{[,VizList,ANY,ANY,ANY-method}
\alias{length,VizList-method}
\alias{names,VizList-method}
\title{DaparVizData class definition}
\usage{
\S4method{Convert2VizList}{QFeatures}(object)

VizList(ll.vizData = NULL)

\S4method{initialize}{VizList}(.Object, ll.vizData)

VizList(ll.vizData = NULL)

\S4method{[}{VizList,ANY,ANY,ANY}(x, i, j = NA, k = NA, l = NA, ..., drop)

\S4method{length}{VizList}(x)

\S4method{names}{VizList}(x)
}
\arguments{
\item{object}{An instance of class \code{QFeatures}.}

\item{ll.vizData}{xxx}

\item{.Object}{xxx}

\item{x}{xxx}

\item{i}{xxx}

\item{j}{xxx}

\item{k}{xxx}

\item{l}{xxx}

\item{...}{xxx}

\item{drop}{xxx}
}
\value{
An instance of 'VizList' class
}
\description{
xxx

xxx

xxx

xxx
}
\section{Slots}{

\describe{
\item{\code{ll.vizData}}{xxx}
}}

\examples{
library(DaparViz)
library(Matrix)


# Convert a QFeatures object to an instance of VizList class
data(ft1)
vData_ft1 <- convert2viz(ft1)
vData_ft1

# Convert a list of MSnSet objects to an instance of VizList class
data(ms1)
data(ms2)
ll.tmp <- setNames(c(ms1, ms2), nm = c('ms1', 'ms2'))

convert2viz(ll.tmp)


# Convert a list of lists of specific items to an instance of VizList class

data(ms1)
msnset <- ms1
X <- PSMatch::makeAdjacencyMatrix(MSnbase::fData(msnset)[, msnset@experimentData@other$proteinId])
rownames(X) <- rownames(MSnbase::fData(msnset))
connectedComp <- PSMatch::ConnectedComponents(X)

ll.ms1 <- list(
  qdata = MSnbase::exprs(msnset),
  metacell = MSnbase::fData(msnset)[ , msnset@experimentData@other$names_metacell],
  mdata = MSnbase::fData(msnset),
  colID = msnset@experimentData@other$keyId,
  conds = MSnbase::pData(msnset)[, 'Condition'],
  type = msnset@experimentData@other$typeOfData,
  adjMat = as.matrix(X),
  cc = as.list(connectedComp@adjMatrices)
  )

data(ms2)
msnset <- ms2
ll.ms2 <- list(
   qdata = MSnbase::exprs(msnset),
  metacell = MSnbase::fData(msnset)[ , msnset@experimentData@other$names_metacell],
  mdata = MSnbase::fData(msnset),
  colID = msnset@experimentData@other$keyId,
  conds = MSnbase::pData(msnset)[, 'Condition'],
  type = msnset@experimentData@other$typeOfData
)


ll.tmp <- list(ms1 = ll.ms1,
               ms2 = ll.ms2)

convert2viz(ll.tmp)
}
